refactor(tests): split html.test.ts into focused template test files
The 4,410-line test/lib/html.test.ts violated multiple test quality
criteria: tautological constant-value assertions, duplicate tests
already covered in attendee-table.test.ts and detail-rows.test.ts,
and poor discoverability due to mixing unrelated concerns.

Changes:
- Delete test/lib/html.test.ts
- Add test/templates/layout.test.ts (asset-paths behavior, nav, login, read-only)
- Add test/templates/payment.test.ts (paymentPage, successPage, cancelPage, etc.)
- Add test/templates/csv.test.ts (generateAttendeesCsv with all variants)
- Add test/templates/public.test.ts (ticketPage, buildOgTags, renderEventImage)
- Add test/templates/admin/dashboard.test.ts (adminDashboardPage, stats, multi-booking)
- Add test/templates/admin/activity-log.test.ts (global and event activity logs)
- Add test/templates/admin/sessions.test.ts (adminSessionsPage)
- Add test/templates/admin/users.test.ts (adminUsersPage, adminUserDeletePage)
- Add test/templates/admin/settings.test.ts (settings + advanced settings pages)
- Add test/templates/admin/questions.test.ts (question CRUD + buildAnswerSummaryRows)
- Add test/templates/admin/calendar.test.ts (calendar page + generateCalendarCsv)
- Add test/templates/admin/events.test.ts (adminEventPage + nearCapacity + isIncompletePayment)
- Split test/lib/forms/ tests out of html.test.ts into focused files

Removed violations:
- Tautological tests asserting CSS_PATH === "/mvp.css" etc. (testing constants, not behavior)
- Duplicate calculateTotalRevenue/countCheckedIn/countCheckedInRows tests (covered in detail-rows.test.ts)
- Duplicate formatAddressInline tests (covered in attendee-table.test.ts)
