+++ test/lib/forms/event-fields.test.ts
@@ -10,7 +10,9 @@ import {
 } from "#templates/fields.ts";
 import { baseEventForm, expectInvalid, expectValid } from "#test-utils";
 
-const eventForm = (overrides: Record<string, string> = {}): Record<string, string> => ({
+const eventForm = (
+  overrides: Record<string, string> = {},
+): Record<string, string> => ({
   ...baseEventForm,
   ...overrides,
 });
@@ -24,14 +26,19 @@ describe("eventFields — required fields", () => {
 
 describe("eventFields — description", () => {
   test("rejects description exceeding max length", () => {
-    expectInvalid(`Description must be ${MAX_TEXTAREA_LENGTH} characters or fewer`)(
+    expectInvalid(
+      `Description must be ${MAX_TEXTAREA_LENGTH} characters or fewer`,
+    )(
       eventFields,
       eventForm({ description: "a".repeat(MAX_TEXTAREA_LENGTH + 1) }),
     );
   });
 
   test("accepts description at and below max length", () => {
-    expectValid(eventFields, eventForm({ description: "a".repeat(MAX_TEXTAREA_LENGTH) }));
+    expectValid(
+      eventFields,
+      eventForm({ description: "a".repeat(MAX_TEXTAREA_LENGTH) }),
+    );
     expectValid(eventFields, eventForm({ description: "" }));
   });
 });
@@ -42,22 +49,37 @@ describe("eventFields — thank_you_url", () => {
   });
 
   test("rejects http and javascript protocols", () => {
-    expectInvalid("URL must use https://")(eventFields, eventForm({ thank_you_url: "http://example.com" }));
-    expectInvalid("URL must use https://")(eventFields, eventForm({ thank_you_url: "javascript:alert(1)" }));
+    expectInvalid("URL must use https://")(
+      eventFields,
+      eventForm({ thank_you_url: "http://example.com" }),
+    );
+    expectInvalid("URL must use https://")(
+      eventFields,
+      eventForm({ thank_you_url: "javascript:alert(1)" }),
+    );
   });
 
   test("rejects invalid URL format", () => {
-    expectInvalid("Invalid URL format")(eventFields, eventForm({ thank_you_url: "not-a-valid-url" }));
+    expectInvalid("Invalid URL format")(
+      eventFields,
+      eventForm({ thank_you_url: "not-a-valid-url" }),
+    );
   });
 });
 
 describe("eventFields — pricing", () => {
   test("rejects negative unit_price", () => {
-    expectInvalid("Price must be 0 or greater")(eventFields, eventForm({ unit_price: "-100" }));
+    expectInvalid("Price must be 0 or greater")(
+      eventFields,
+      eventForm({ unit_price: "-100" }),
+    );
   });
 
   test("rejects negative max_price and accepts valid values", () => {
-    expectInvalid("Price must be 0 or greater")(eventFields, eventForm({ max_price: "-50" }));
+    expectInvalid("Price must be 0 or greater")(
+      eventFields,
+      eventForm({ max_price: "-50" }),
+    );
     expectValid(eventFields, eventForm({ max_price: "100.00" }));
     expectValid(eventFields, eventForm({ max_price: "" }));
   });
@@ -65,11 +87,20 @@ describe("eventFields — pricing", () => {
 
 describe("eventFields — contact fields setting", () => {
   test("rejects unknown contact field name", () => {
-    expectInvalid("Invalid contact field: invalid")(eventFields, eventForm({ fields: "invalid" }));
+    expectInvalid("Invalid contact field: invalid")(
+      eventFields,
+      eventForm({ fields: "invalid" }),
+    );
   });
 
   test("accepts known contact field values", () => {
-    for (const value of ["email", "phone", "address", "special_instructions", "email,phone"]) {
+    for (const value of [
+      "email",
+      "phone",
+      "address",
+      "special_instructions",
+      "email,phone",
+    ]) {
       expectValid(eventFields, eventForm({ fields: value }));
     }
   });
@@ -92,8 +123,17 @@ describe("eventFields — event_type", () => {
 
 describe("eventFields — bookable_days", () => {
   test("accepts valid day names", () => {
-    expectValid(eventFields, eventForm({ bookable_days: "Monday,Wednesday,Friday" }));
-    expectValid(eventFields, eventForm({ bookable_days: "Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday" }));
+    expectValid(
+      eventFields,
+      eventForm({ bookable_days: "Monday,Wednesday,Friday" }),
+    );
+    expectValid(
+      eventFields,
+      eventForm({
+        bookable_days:
+          "Monday,Tuesday,Wednesday,Thursday,Friday,Saturday,Sunday",
+      }),
+    );
   });
 
   test("rejects invalid day name", () => {
@@ -103,7 +143,10 @@ describe("eventFields — bookable_days", () => {
   });
 
   test("rejects empty-after-trimming value", () => {
-    expectInvalid("At least one day is required")(eventFields, eventForm({ bookable_days: "," }));
+    expectInvalid("At least one day is required")(
+      eventFields,
+      eventForm({ bookable_days: "," }),
+    );
   });
 
   test("accepts empty value (optional field)", () => {
@@ -113,20 +156,26 @@ describe("eventFields — bookable_days", () => {
 
 describe("groupCreateFields", () => {
   test("rejects terms_and_conditions exceeding MAX_TEXTAREA_LENGTH", () => {
-    const termsField = groupCreateFields.find((f) => f.name === "terms_and_conditions")!;
-    expect(termsField.validate?.("x".repeat(MAX_TEXTAREA_LENGTH + 1))).toContain(
-      `${MAX_TEXTAREA_LENGTH} characters or fewer`,
-    );
+    const termsField = groupCreateFields.find(
+      (f) => f.name === "terms_and_conditions",
+    )!;
+    expect(
+      termsField.validate?.("x".repeat(MAX_TEXTAREA_LENGTH + 1)),
+    ).toContain(`${MAX_TEXTAREA_LENGTH} characters or fewer`);
   });
 
   test("accepts terms_and_conditions within MAX_TEXTAREA_LENGTH", () => {
-    const termsField = groupCreateFields.find((f) => f.name === "terms_and_conditions")!;
+    const termsField = groupCreateFields.find(
+      (f) => f.name === "terms_and_conditions",
+    )!;
     expect(termsField.validate?.("x".repeat(MAX_TEXTAREA_LENGTH))).toBeNull();
   });
 });
 
 describe("holidayFields", () => {
-  const holidayForm = (overrides: Record<string, string> = {}): Record<string, string> => ({
+  const holidayForm = (
+    overrides: Record<string, string> = {},
+  ): Record<string, string> => ({
     name: "Bank Holiday",
     start_date: "2026-12-25",
     end_date: "2026-12-25",
@@ -134,21 +183,39 @@ describe("holidayFields", () => {
   });
 
   test("rejects missing name, start_date, or end_date", () => {
-    expectInvalid("Holiday Name is required")(holidayFields, holidayForm({ name: "" }));
-    expectInvalid("Start Date is required")(holidayFields, holidayForm({ start_date: "" }));
-    expectInvalid("End Date is required")(holidayFields, holidayForm({ end_date: "" }));
+    expectInvalid("Holiday Name is required")(
+      holidayFields,
+      holidayForm({ name: "" }),
+    );
+    expectInvalid("Start Date is required")(
+      holidayFields,
+      holidayForm({ start_date: "" }),
+    );
+    expectInvalid("End Date is required")(
+      holidayFields,
+      holidayForm({ end_date: "" }),
+    );
   });
 
   test("rejects malformed dates", () => {
-    expectInvalid("Please enter a valid date (YYYY-MM-DD)")(holidayFields, holidayForm({ start_date: "25-12-2026" }));
-    expectInvalid("Please enter a valid date (YYYY-MM-DD)")(holidayFields, holidayForm({ end_date: "not-a-date" }));
+    expectInvalid("Please enter a valid date (YYYY-MM-DD)")(
+      holidayFields,
+      holidayForm({ start_date: "25-12-2026" }),
+    );
+    expectInvalid("Please enter a valid date (YYYY-MM-DD)")(
+      holidayFields,
+      holidayForm({ end_date: "not-a-date" }),
+    );
   });
 
   test("accepts valid single-day and multi-day holidays", () => {
     const values = expectValid(holidayFields, holidayForm());
     expect(values.name).toBe("Bank Holiday");
 
-    expectValid(holidayFields, holidayForm({ start_date: "2026-12-24", end_date: "2026-12-26" }));
+    expectValid(
+      holidayFields,
+      holidayForm({ start_date: "2026-12-24", end_date: "2026-12-26" }),
+    );
   });
 });
 
@@ -159,9 +226,15 @@ describe("validateDate", () => {
   });
 
   test("rejects wrong format", () => {
-    expect(validateDate("12/25/2026")).toBe("Please enter a valid date (YYYY-MM-DD)");
-    expect(validateDate("2026-12")).toBe("Please enter a valid date (YYYY-MM-DD)");
-    expect(validateDate("not-a-date")).toBe("Please enter a valid date (YYYY-MM-DD)");
+    expect(validateDate("12/25/2026")).toBe(
+      "Please enter a valid date (YYYY-MM-DD)",
+    );
+    expect(validateDate("2026-12")).toBe(
+      "Please enter a valid date (YYYY-MM-DD)",
+    );
+    expect(validateDate("not-a-date")).toBe(
+      "Please enter a valid date (YYYY-MM-DD)",
+    );
     expect(validateDate("")).toBe("Please enter a valid date (YYYY-MM-DD)");
   });
 
@@ -179,7 +252,9 @@ describe("validateBookableDays", () => {
   });
 
   test("rejects invalid day name", () => {
-    expect(validateBookableDays("Monday,Funday")).toContain("Invalid day: Funday");
+    expect(validateBookableDays("Monday,Funday")).toContain(
+      "Invalid day: Funday",
+    );
   });
 
   test("rejects empty or blank-only values", () => {
