+++ test/lib/forms/rendering.test.ts
@@ -1,5 +1,6 @@
 import { expect } from "@std/expect";
 import { afterEach, describe, it as test } from "@std/testing/bdd";
+import { FormParams } from "#lib/form-data.ts";
 import {
   clearSavedFormData,
   type Field,
@@ -7,7 +8,6 @@ import {
   renderFields,
   setSavedFormData,
 } from "#lib/forms.tsx";
-import { FormParams } from "#lib/form-data.ts";
 
 const field = (
   overrides: Partial<Field> & { name: string; label: string },
@@ -28,29 +28,52 @@ describe("renderField", () => {
   });
 
   test("renders required attribute", () => {
-    const html = rendered({ name: "email", label: "Email", type: "email", required: true });
+    const html = rendered({
+      name: "email",
+      label: "Email",
+      type: "email",
+      required: true,
+    });
     expect(html).toContain("required");
   });
 
   test("renders placeholder", () => {
-    const html = rendered({ name: "name", label: "Name", placeholder: "Enter your name" });
+    const html = rendered({
+      name: "name",
+      label: "Name",
+      placeholder: "Enter your name",
+    });
     expect(html).toContain('placeholder="Enter your name"');
   });
 
   test("renders hint text", () => {
-    const html = rendered({ name: "pw", label: "Password", type: "password", hint: "Minimum 8 characters" });
+    const html = rendered({
+      name: "pw",
+      label: "Password",
+      type: "password",
+      hint: "Minimum 8 characters",
+    });
     expect(html).toContain("Minimum 8 characters");
     expect(html).toContain("<small");
   });
 
   test("renders hintHtml as raw HTML", () => {
-    const html = rendered({ name: "desc", label: "Description", hintHtml: '<a href="/guide">Help</a>' });
+    const html = rendered({
+      name: "desc",
+      label: "Description",
+      hintHtml: '<a href="/guide">Help</a>',
+    });
     expect(html).toContain('<a href="/guide">Help</a>');
     expect(html).toContain("<small");
   });
 
   test("renders min attribute for number", () => {
-    const html = rendered({ name: "qty", label: "Quantity", type: "number", min: 1 });
+    const html = rendered({
+      name: "qty",
+      label: "Quantity",
+      type: "number",
+      min: 1,
+    });
     expect(html).toContain('min="1"');
   });
 
@@ -60,33 +83,53 @@ describe("renderField", () => {
   });
 
   test("renders maxlength attribute", () => {
-    const html = rendered({ name: "desc", label: "Description", maxlength: 128 });
+    const html = rendered({
+      name: "desc",
+      label: "Description",
+      maxlength: 128,
+    });
     expect(html).toContain('maxlength="128"');
   });
 
   test("renders autocomplete when provided, omits when absent", () => {
-    expect(rendered({ name: "name", label: "Name", autocomplete: "name" })).toContain('autocomplete="name"');
-    expect(rendered({ name: "name", label: "Name" })).not.toContain("autocomplete");
+    expect(
+      rendered({ name: "name", label: "Name", autocomplete: "name" }),
+    ).toContain('autocomplete="name"');
+    expect(rendered({ name: "name", label: "Name" })).not.toContain(
+      "autocomplete",
+    );
   });
 
   test("renders textarea for textarea type", () => {
-    const html = rendered({ name: "description", label: "Description", type: "textarea" });
+    const html = rendered({
+      name: "description",
+      label: "Description",
+      type: "textarea",
+    });
     expect(html).toContain("<textarea");
     expect(html).not.toContain("<input");
   });
 
   test("renders value when provided", () => {
-    expect(rendered({ name: "name", label: "Name" }, "John")).toContain('value="John"');
+    expect(rendered({ name: "name", label: "Name" }, "John")).toContain(
+      'value="John"',
+    );
   });
 
   test("escapes HTML in value", () => {
-    const html = rendered({ name: "name", label: "Name" }, '<script>alert("xss")</script>');
+    const html = rendered(
+      { name: "name", label: "Name" },
+      '<script>alert("xss")</script>',
+    );
     expect(html).toContain("&lt;script&gt;");
     expect(html).not.toContain("<script>");
   });
 
   test("renders textarea with value", () => {
-    const html = rendered({ name: "description", label: "Description", type: "textarea" }, "Some description");
+    const html = rendered(
+      { name: "description", label: "Description", type: "textarea" },
+      "Some description",
+    );
     expect(html).toContain(">Some description</textarea>");
   });
 
@@ -129,20 +172,31 @@ describe("renderField", () => {
 
   describe("date type", () => {
     test("renders date input", () => {
-      const html = rendered({ name: "start_date", label: "Start Date", type: "date" });
+      const html = rendered({
+        name: "start_date",
+        label: "Start Date",
+        type: "date",
+      });
       expect(html).toContain('type="date"');
       expect(html).toContain('name="start_date"');
     });
 
     test("renders date input with value", () => {
-      const html = rendered({ name: "start_date", label: "Start Date", type: "date" }, "2026-12-25");
+      const html = rendered(
+        { name: "start_date", label: "Start Date", type: "date" },
+        "2026-12-25",
+      );
       expect(html).toContain('value="2026-12-25"');
     });
   });
 
   describe("datetime type", () => {
     test("renders split date and time inputs", () => {
-      const html = rendered({ name: "closes_at", label: "Closes At", type: "datetime" });
+      const html = rendered({
+        name: "closes_at",
+        label: "Closes At",
+        type: "datetime",
+      });
       expect(html).toContain('name="closes_at_date"');
       expect(html).toContain('placeholder="Date"');
       expect(html).toContain('name="closes_at_time"');
@@ -150,20 +204,31 @@ describe("renderField", () => {
     });
 
     test("splits combined value into date and time parts", () => {
-      const html = rendered({ name: "closes_at", label: "Closes At", type: "datetime" }, "2099-06-15T14:30");
+      const html = rendered(
+        { name: "closes_at", label: "Closes At", type: "datetime" },
+        "2099-06-15T14:30",
+      );
       expect(html).toContain('value="2099-06-15"');
       expect(html).toContain('value="14:30"');
     });
 
     test("renders no value attributes when value is empty", () => {
-      const html = rendered({ name: "closes_at", label: "Closes At", type: "datetime" }, "");
+      const html = rendered(
+        { name: "closes_at", label: "Closes At", type: "datetime" },
+        "",
+      );
       expect(html).not.toContain("value=");
     });
   });
 
   describe("file type", () => {
     test("renders file input with accept attribute", () => {
-      const html = rendered({ name: "image", label: "Upload Image", type: "file", accept: "image/jpeg,image/png" });
+      const html = rendered({
+        name: "image",
+        label: "Upload Image",
+        type: "file",
+        accept: "image/jpeg,image/png",
+      });
       expect(html).toContain('type="file"');
       expect(html).toContain('accept="image/jpeg,image/png"');
     });
@@ -224,12 +289,17 @@ describe("renderFields", () => {
   });
 
   test("omits value attribute for null values", () => {
-    const html = renderFields([field({ name: "price", label: "Price", type: "number" })], { price: null });
+    const html = renderFields(
+      [field({ name: "price", label: "Price", type: "number" })],
+      { price: null },
+    );
     expect(html).not.toContain('value="null"');
   });
 
   test("uses defaultValue when no explicit value or saved data", () => {
-    const html = renderFields([field({ name: "country", label: "Country", defaultValue: "US" })]);
+    const html = renderFields([
+      field({ name: "country", label: "Country", defaultValue: "US" }),
+    ]);
     expect(html).toContain('value="US"');
   });
 });
@@ -239,19 +309,25 @@ describe("renderFields with saved form data", () => {
 
   test("restores saved text values", () => {
     setSavedFormData(new FormParams("name=Alice"));
-    expect(renderFields([field({ name: "name", label: "Name" })])).toContain('value="Alice"');
+    expect(renderFields([field({ name: "name", label: "Name" })])).toContain(
+      'value="Alice"',
+    );
   });
 
   test("explicit values take precedence over saved values", () => {
     setSavedFormData(new FormParams("name=Saved"));
-    const html = renderFields([field({ name: "name", label: "Name" })], { name: "Explicit" });
+    const html = renderFields([field({ name: "name", label: "Name" })], {
+      name: "Explicit",
+    });
     expect(html).toContain('value="Explicit"');
     expect(html).not.toContain("Saved");
   });
 
   test("saved data takes precedence over defaultValue", () => {
     setSavedFormData(new FormParams("country=GB"));
-    const html = renderFields([field({ name: "country", label: "Country", defaultValue: "US" })]);
+    const html = renderFields([
+      field({ name: "country", label: "Country", defaultValue: "US" }),
+    ]);
     expect(html).toContain('value="GB"');
     expect(html).not.toContain('value="US"');
   });
