--- test/lib/html.test.ts
@@ -1,4410 +1,0 @@
-import { expect } from "@std/expect";
-import { afterEach, beforeAll, describe, it as test } from "@std/testing/bdd";
-import {
-  CSS_PATH,
-  EMBED_JS_PATH,
-  IFRAME_RESIZER_CHILD_JS_PATH,
-  IFRAME_RESIZER_PARENT_JS_PATH,
-  JS_PATH,
-} from "#lib/asset-paths.ts";
-import { getCurrentCsrfToken, signCsrfToken } from "#lib/csrf.ts";
-import { addDays } from "#lib/dates.ts";
-import { settings } from "#lib/db/settings.ts";
-import { detectIframeMode } from "#lib/iframe.ts";
-import { runWithStorageConfig } from "#lib/storage.ts";
-import { todayInTz } from "#lib/timezone.ts";
-import type { EventWithCount } from "#lib/types.ts";
-import {
-  adminEventActivityLogPage,
-  adminGlobalActivityLogPage,
-} from "#templates/admin/activityLog.tsx";
-import {
-  adminCalendarPage,
-  type CalendarAttendeeRow,
-} from "#templates/admin/calendar.tsx";
-import {
-  activeEventStatsSection,
-  adminDashboardPage,
-} from "#templates/admin/dashboard.tsx";
-import {
-  adminDuplicateEventPage,
-  adminEventEditPage,
-  adminEventNewPage,
-  adminEventPage,
-  buildAnswerSummaryRows,
-  calculateTotalRevenue,
-  countCheckedIn,
-  countCheckedInRows,
-  formatAddressInline,
-  isIncompletePayment,
-  nearCapacity,
-} from "#templates/admin/events.tsx";
-import { adminLoginPage } from "#templates/admin/login.tsx";
-import { AdminNav } from "#templates/admin/nav.tsx";
-import {
-  adminAnswerDeletePage,
-  adminEventQuestionsPage,
-  adminQuestionDeletePage,
-  adminQuestionPage,
-  adminQuestionsPage,
-} from "#templates/admin/questions.tsx";
-import { adminSessionsPage } from "#templates/admin/sessions.tsx";
-import { adminSettingsPage } from "#templates/admin/settings.tsx";
-import { adminAdvancedSettingsPage } from "#templates/admin/settings-advanced.tsx";
-import {
-  adminUserDeletePage,
-  adminUsersPage,
-  type DisplayUser,
-} from "#templates/admin/users.tsx";
-import {
-  type CsvEventInfo,
-  generateAttendeesCsv,
-  generateCalendarCsv,
-} from "#templates/csv.ts";
-import { eventFields } from "#templates/fields.ts";
-import { Layout } from "#templates/layout.tsx";
-import {
-  checkoutPopupPage,
-  paymentCancelPage,
-  paymentErrorPage,
-  paymentPage,
-  successPage,
-} from "#templates/payment.tsx";
-import {
-  buildOgTags,
-  buildTicketEvent,
-  notFoundPage,
-  renderEventImage,
-  temporaryErrorPage,
-  ticketPage,
-} from "#templates/public.tsx";
-import { ticketViewPage } from "#templates/tickets.tsx";
-import {
-  describeWithEnv,
-  setupTestEncryptionKey,
-  testAttendee,
-  testEvent,
-  testEventWithCount,
-  testGroup,
-  withStorageDisabled,
-} from "#test-utils";
-
-const TEST_SESSION = { adminLevel: "owner" as const };
-
-beforeAll(async () => {
-  setupTestEncryptionKey();
-  await signCsrfToken();
-});
-
-describe("asset-paths", () => {
-  test("CSS_PATH defaults to /mvp.css in dev", () => {
-    expect(CSS_PATH).toBe("/mvp.css");
-  });
-
-  test("pages include CSS_PATH in stylesheet link", () => {
-    const html = adminLoginPage();
-    expect(html).toContain(`href="${CSS_PATH}"`);
-    expect(html).toContain('rel="stylesheet"');
-  });
-
-  test("JS_PATH defaults to /admin.js in dev", () => {
-    expect(JS_PATH).toBe("/admin.js");
-  });
-
-  test("pages include JS_PATH in deferred script tag", () => {
-    const html = adminLoginPage();
-    expect(html).toContain(`src="${JS_PATH}"`);
-    expect(html).toContain("defer");
-  });
-
-  test("IFRAME_RESIZER_PARENT_JS_PATH defaults to /iframe-resizer-parent.js in dev", () => {
-    expect(IFRAME_RESIZER_PARENT_JS_PATH).toBe("/iframe-resizer-parent.js");
-  });
-
-  test("EMBED_JS_PATH defaults to /embed.js in dev", () => {
-    expect(EMBED_JS_PATH).toBe("/embed.js");
-  });
-
-  test("IFRAME_RESIZER_CHILD_JS_PATH defaults to /iframe-resizer-child.js in dev", () => {
-    expect(IFRAME_RESIZER_CHILD_JS_PATH).toBe("/iframe-resizer-child.js");
-  });
-
-  test("EMBED_JS_PATH defaults to /embed.js in dev", () => {
-    expect(EMBED_JS_PATH).toBe("/embed.js");
-  });
-});
-
-describe("html", () => {
-  afterEach(() => {
-    detectIframeMode("https://example.com/");
-  });
-
-  describe("Layout skip navigation", () => {
-    test("renders skip-nav link targeting main-content", () => {
-      const html = String(Layout({ title: "Test", children: "" }));
-      expect(html).toContain('class="skip-nav"');
-      expect(html).toContain('href="#main-content"');
-      expect(html).toContain("Skip to content");
-      expect(html).toContain('id="main-content"');
-      expect(html).toContain('tabindex="-1"');
-    });
-  });
-
-  describe("adminLoginPage", () => {
-    test("renders login form", () => {
-      const html = adminLoginPage();
-      expect(html).toContain("Login");
-      expect(html).toContain('action="/admin/login"');
-      expect(html).toContain('type="password"');
-      expect(html).toContain('name="csrf_token"');
-    });
-
-    test("shows error when provided", () => {
-      const html = adminLoginPage("Invalid password");
-      expect(html).toContain("Invalid password");
-      expect(html).toContain('class="error"');
-    });
-
-    test("escapes error message", () => {
-      const html = adminLoginPage("<script>evil()</script>");
-      expect(html).toContain("&lt;script&gt;");
-    });
-  });
-
-  describe("adminDashboardPage", () => {
-    test("renders empty state when no events", () => {
-      const html = adminDashboardPage([], TEST_SESSION);
-      expect(html).toContain("Events");
-      expect(html).toContain("No events yet");
-    });
-
-    test("renders events table", () => {
-      const events = [testEventWithCount({ attendee_count: 25 })];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("Test Event");
-      expect(html).toContain("25 / 100");
-      expect(html).toContain("/admin/event/1");
-    });
-
-    test("displays event name", () => {
-      const events = [testEventWithCount({ name: "My Test Event" })];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("My Test Event");
-      expect(html).toContain("Event Name");
-    });
-
-    test("renders add event link", () => {
-      const html = adminDashboardPage([], TEST_SESSION);
-      expect(html).toContain('href="/admin/event/new"');
-      expect(html).toContain("Add Event");
-    });
-
-    test("includes logout link", () => {
-      const html = adminDashboardPage([], TEST_SESSION);
-      expect(html).toContain("/admin/logout");
-    });
-
-    test("renders newest attendees in an open details element", () => {
-      const events = [testEventWithCount({ id: 1, name: "Gala Night" })];
-      const attendees = [
-        testAttendee({ id: 1, event_id: 1, name: "Alice" }),
-        testAttendee({ id: 2, event_id: 1, name: "Bob" }),
-      ];
-      const html = adminDashboardPage(
-        events,
-        TEST_SESSION,
-        undefined,
-        attendees,
-      );
-      expect(html).toContain("<details open");
-      expect(html).toContain("Newest 2 Attendees");
-    });
-
-    test("newest attendees section not shown when no attendees", () => {
-      const html = adminDashboardPage([], TEST_SESSION, undefined, []);
-      expect(html).not.toContain("Newest");
-      expect(html).not.toContain("<details open");
-    });
-
-    test("newest attendees shows singular for single attendee", () => {
-      const events = [testEventWithCount({ id: 1 })];
-      const attendees = [testAttendee({ id: 1, event_id: 1 })];
-      const html = adminDashboardPage(
-        events,
-        TEST_SESSION,
-        undefined,
-        attendees,
-      );
-      expect(html).toContain("Newest 1 Attendee</summary>");
-    });
-
-    test("newest attendees shows event column", () => {
-      const events = [testEventWithCount({ id: 1, name: "Workshop" })];
-      const attendees = [testAttendee({ id: 1, event_id: 1 })];
-      const html = adminDashboardPage(
-        events,
-        TEST_SESSION,
-        undefined,
-        attendees,
-      );
-      expect(html).toContain("<th>Event</th>");
-      expect(html).toContain("Workshop");
-    });
-
-    test("newest attendees not shown when all attendees have unknown event_id", () => {
-      const events = [testEventWithCount({ id: 1 })];
-      const attendees = [testAttendee({ id: 1, event_id: 999 })];
-      const html = adminDashboardPage(
-        events,
-        TEST_SESSION,
-        undefined,
-        attendees,
-      );
-      expect(html).not.toContain("Newest");
-      expect(html).not.toContain("<details open");
-    });
-
-    test("newest attendees skips attendees with unknown event_id", () => {
-      const events = [testEventWithCount({ id: 1, name: "Known Event" })];
-      const attendees = [
-        testAttendee({ id: 1, event_id: 1, name: "Valid" }),
-        testAttendee({ id: 2, event_id: 999, name: "Orphan" }),
-      ];
-      const html = adminDashboardPage(
-        events,
-        TEST_SESSION,
-        undefined,
-        attendees,
-      );
-      expect(html).toContain("Valid");
-      expect(html).not.toContain("Orphan");
-      expect(html).toContain("Newest 1 Attendee</summary>");
-    });
-  });
-
-  describe("adminEventEditPage group select", () => {
-    test("preselects the event group_id when groups exist", () => {
-      const groups = [testGroup({ id: 2, name: "Group Two" })];
-      const event = testEventWithCount({ group_id: 2 });
-      const html = adminEventEditPage(event, groups, TEST_SESSION);
-      expect(html).toContain('name="group_id"');
-      expect(html).toContain('<option value="2" selected');
-      expect(html).not.toContain('<option value="0" selected');
-    });
-  });
-
-  describe("adminEventPage", () => {
-    const event = testEventWithCount({ attendee_count: 2 });
-
-    test("renders event name", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Test Event");
-    });
-
-    test("shows attendees row with count and remaining", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Attendees");
-      expect(html).toContain("2 / 100");
-      expect(html).toContain("98 remain");
-    });
-
-    test("shows checked in row with 0 of 0 when no attendees", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Checked In");
-      expect(html).toContain("0 / 0");
-      expect(html).toContain("0 remain");
-    });
-
-    test("shows checked in count and remaining", () => {
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true }),
-        testAttendee({ id: 2, checked_in: false }),
-        testAttendee({ id: 3, checked_in: false }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Checked In");
-      expect(html).toContain("1 / 3");
-      expect(html).toContain("2 remain");
-    });
-
-    test("shows dual checked-in rows when attendees have multi-quantity", () => {
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true, quantity: 2 }),
-        testAttendee({ id: 2, checked_in: false, quantity: 3 }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      // Tickets Checked In: 1 row / 2 rows, 1 remain
-      expect(html).toContain("Tickets Checked In");
-      expect(html).toContain("1 / 2");
-      expect(html).toContain("1 remain");
-      // Attendees Checked In: 2 qty / 5 total qty, 3 remain
-      expect(html).toContain("Attendees Checked In");
-      expect(html).toContain("2 / 5");
-      expect(html).toContain("3 remain");
-    });
-
-    test("dual checked-in rows show daily suffix when daily event with dateFilter", () => {
-      const dailyEvent = testEventWithCount({
-        event_type: "daily",
-        attendee_count: 5,
-      });
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true, quantity: 2 }),
-        testAttendee({ id: 2, checked_in: false, quantity: 3 }),
-      ];
-      const html = adminEventPage({
-        event: dailyEvent,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        dateFilter: "2026-03-15",
-      });
-      expect(html).toContain("Attendees Checked In (Sunday 15 March 2026)");
-      expect(html).toContain("Tickets Checked In (Sunday 15 March 2026)");
-    });
-
-    test("dual checked-in rows show total suffix when daily event without dateFilter", () => {
-      const dailyEvent = testEventWithCount({
-        event_type: "daily",
-        attendee_count: 5,
-      });
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true, quantity: 2 }),
-        testAttendee({ id: 2, checked_in: false, quantity: 3 }),
-      ];
-      const html = adminEventPage({
-        event: dailyEvent,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Attendees Checked In (total)");
-      expect(html).toContain("Tickets Checked In (total)");
-    });
-
-    test("shows thank you URL in copyable input", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Thank You URL");
-      expect(html).toContain('value="https://example.com/thanks"');
-      expect(html).toContain("readonly");
-      expect(html).toContain("data-select-on-click");
-    });
-
-    test("shows public URL with allowed domain", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "example.com",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Public URL");
-      expect(html).toContain('href="https://example.com/ticket/ab12c"');
-      expect(html).toContain("example.com/ticket/ab12c");
-    });
-
-    test("shows embed codes with allowed domain and iframe param", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "example.com",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Embed Script");
-      expect(html).toContain("Embed Iframe");
-      expect(html).toContain("embed.js");
-      expect(html).toContain("data-events=");
-      expect(html).toContain("https://example.com/ticket/ab12c?iframe=true");
-      expect(html).toContain("height: 600px");
-      expect(html).toContain("loading=");
-      expect(html).toContain("readonly");
-    });
-
-    test("iframe embed is a plain iframe without resizer scripts", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "example.com",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("iframeResize");
-    });
-
-    test("renders empty attendees state", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("No attendees yet");
-    });
-
-    test("renders attendees table", () => {
-      const attendees = [testAttendee()];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("John Doe");
-      expect(html).toContain("john@example.com");
-    });
-
-    test("escapes attendee data", () => {
-      const attendees = [testAttendee({ name: "<script>evil()</script>" })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("&lt;script&gt;");
-    });
-
-    test("includes back link", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("/admin/");
-    });
-
-    test("shows phone column when attendee has phone", () => {
-      const attendees = [testAttendee({ phone: "+1 555 123 4567" })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("<th>Phone</th>");
-      expect(html).toContain("+1 555 123 4567");
-    });
-
-    test("hides phone column when no attendees have phone", () => {
-      const attendees = [testAttendee({ phone: "" })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("<th>Phone</th>");
-    });
-
-    test("hides email column when no attendees have email", () => {
-      const attendees = [testAttendee({ email: "" })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("John Doe");
-      expect(html).not.toContain("<th>Email</th>");
-    });
-
-    test("shows danger-text class when near capacity", () => {
-      const nearFullEvent = testEventWithCount({
-        attendee_count: 91,
-        max_attendees: 100,
-      });
-      const html = adminEventPage({
-        event: nearFullEvent,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain('class="danger-text"');
-      expect(html).toContain("9 remain");
-    });
-
-    test("does not show danger-text class when not near capacity", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain('class="danger-text"');
-    });
-
-    test("shows deactivated alert for inactive events", () => {
-      const inactive = testEventWithCount({ active: false, attendee_count: 0 });
-      const html = adminEventPage({
-        event: inactive,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain('class="error"');
-      expect(html).toContain("This event is deactivated and cannot be booked");
-    });
-
-    test("does not show deactivated alert for active events", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain(
-        "This event is deactivated and cannot be booked",
-      );
-    });
-
-    test("shows ticket column header", () => {
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("<th>Ticket</th>");
-    });
-
-    test("shows ticket token as link to public ticket URL", () => {
-      const attendees = [testAttendee({ ticket_token: "abc123" })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "mysite.com",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain('href="https://mysite.com/t/abc123"');
-      expect(html).toContain("abc123");
-    });
-
-    test("renders empty date cell for attendee without date on daily event", () => {
-      const dailyEvent = testEventWithCount({
-        event_type: "daily",
-        attendee_count: 1,
-      });
-      const attendees = [testAttendee({ date: null })];
-      const html = adminEventPage({
-        event: dailyEvent,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("<th>Date</th>");
-    });
-
-    test("shows unlimited booking window when maximum_days_after is 0", () => {
-      const dailyEvent = testEventWithCount({
-        event_type: "daily",
-        maximum_days_after: 0,
-      });
-      const html = adminEventPage({
-        event: dailyEvent,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("unlimited");
-    });
-
-    test("shows numeric booking window when maximum_days_after is nonzero", () => {
-      const dailyEvent = testEventWithCount({
-        event_type: "daily",
-        maximum_days_after: 30,
-      });
-      const html = adminEventPage({
-        event: dailyEvent,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("to 30 days");
-      expect(html).not.toContain("unlimited");
-    });
-
-    test("shows danger-text for daily event at capacity with date filter", () => {
-      const dailyEvent = testEventWithCount({
-        event_type: "daily",
-        attendee_count: 0,
-        max_attendees: 2,
-      });
-      const attendees = [testAttendee(), testAttendee({ id: 2, name: "Jane" })];
-      const html = adminEventPage({
-        event: dailyEvent,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        dateFilter: "2026-03-15",
-      });
-      expect(html).toContain('class="danger-text"');
-      expect(html).toContain("0 remain");
-    });
-  });
-
-  describe("ticketPage (single event)", () => {
-    const event = testEventWithCount({ attendee_count: 50 });
-    const renderTicket = (
-      ev: EventWithCount,
-      opts?: {
-        error?: string;
-        isClosed?: boolean;
-        iframe?: boolean;
-        dates?: string[];
-        terms?: string | null;
-        baseUrl?: string;
-        questions?: {
-          id: number;
-          text: string;
-          answers: {
-            id: number;
-            question_id: number;
-            text: string;
-            sort_order: number;
-          }[];
-        }[];
-      },
-    ) => {
-      if (opts?.iframe) detectIframeMode("https://example.com/?iframe=true");
-      else detectIframeMode("https://example.com/");
-      return ticketPage({
-        events: [buildTicketEvent(ev, opts?.isClosed)],
-        slugs: [ev.slug],
-        error: opts?.error,
-        dates: opts?.dates ?? [],
-        terms: opts?.terms,
-        baseUrl: opts?.baseUrl,
-        questions: opts?.questions,
-      });
-    };
-
-    test("renders page title", () => {
-      const html = renderTicket(event);
-      expect(html).toContain("Test Event");
-    });
-
-    test("renders registration form when spots available", () => {
-      const html = renderTicket(event);
-      expect(html).toContain('action="/ticket/ab12c"');
-      expect(html).toContain('name="name"');
-      expect(html).toContain('name="email"');
-      expect(html).toContain("Continue");
-    });
-
-    test("includes CSRF token in form", () => {
-      const html = renderTicket(event);
-      expect(html).toContain('name="csrf_token"');
-      expect(html).toContain(`value="${getCurrentCsrfToken()}"`);
-    });
-
-    test("shows error when provided", () => {
-      const html = renderTicket(event, {
-        error: "Name and email are required",
-      });
-      expect(html).toContain("Name and email are required");
-      expect(html).toContain('class="error"');
-    });
-
-    test("shows full message when no spots", () => {
-      const fullEvent = testEventWithCount({ attendee_count: 100 });
-      const html = renderTicket(fullEvent);
-      expect(html).toContain("this event is full");
-      expect(html).not.toContain(">Reserve Ticket</button>");
-    });
-
-    test("displays event name as header", () => {
-      const html = renderTicket(event);
-      expect(html).toContain("<h1>Test Event</h1>");
-    });
-
-    test("shows quantity selector when max_quantity > 1 and spots available", () => {
-      const multiQtyEvent = testEventWithCount({
-        max_quantity: 5,
-        attendee_count: 0,
-      });
-      const html = renderTicket(multiQtyEvent);
-      expect(html).toContain("Number of Tickets");
-      expect(html).toContain(`name="quantity_${multiQtyEvent.id}"`);
-      expect(html).toContain('<option value="1">1</option>');
-      expect(html).toContain('<option value="5">5</option>');
-      expect(html).toContain("Continue");
-    });
-
-    test("limits quantity selector to remaining spots", () => {
-      const limitedEvent = testEventWithCount({
-        max_quantity: 10,
-        attendee_count: 97, // Only 3 spots remaining
-      });
-      const html = renderTicket(limitedEvent);
-      expect(html).toContain("Number of Tickets");
-      expect(html).toContain('<option value="3">3</option>');
-      expect(html).not.toContain('<option value="4">4</option>');
-    });
-
-    test("hides quantity selector when max_quantity is 1", () => {
-      const html = renderTicket(event); // max_quantity is 1
-      expect(html).not.toContain("Number of Tickets");
-      expect(html).toContain(
-        `type="hidden" name="quantity_${event.id}" value="1"`,
-      );
-      expect(html).toContain("Continue");
-    });
-
-    test("shows Continue button for purchase_only event", () => {
-      const poEvent = testEventWithCount({
-        attendee_count: 50,
-        purchase_only: true,
-      });
-      const html = renderTicket(poEvent);
-      expect(html).toContain("Continue");
-    });
-
-    test("shows phone field for phone-only events", () => {
-      const phoneEvent = testEventWithCount({
-        attendee_count: 50,
-        fields: "phone",
-      });
-      const html = renderTicket(phoneEvent);
-      expect(html).toContain('name="phone"');
-      expect(html).toContain("Your Phone Number");
-      expect(html).not.toContain('name="email"');
-    });
-
-    test("shows both email and phone for email,phone setting", () => {
-      const bothEvent = testEventWithCount({
-        attendee_count: 50,
-        fields: "email,phone",
-      });
-      const html = renderTicket(bothEvent);
-      expect(html).toContain('name="email"');
-      expect(html).toContain('name="phone"');
-    });
-
-    test("shows only email for email setting", () => {
-      const html = renderTicket(event);
-      expect(html).toContain('name="email"');
-      expect(html).not.toContain('name="phone"');
-    });
-
-    test("hides header and description in iframe mode", () => {
-      const eventWithDesc = testEventWithCount({
-        attendee_count: 50,
-        description: "A great event",
-      });
-      const html = renderTicket(eventWithDesc, { iframe: true });
-      expect(html).not.toContain("<h1>");
-      expect(html).not.toContain("A great event");
-      expect(html).toContain('class="iframe"');
-      expect(html).toContain('name="name"');
-    });
-
-    test("shows header and description when not in iframe mode", () => {
-      const eventWithDesc = testEventWithCount({
-        attendee_count: 50,
-        description: "A great event",
-      });
-      const html = renderTicket(eventWithDesc);
-      expect(html).toContain("<h1>Test Event</h1>");
-      expect(html).toContain("A great event");
-      expect(html).not.toContain('class="iframe"');
-    });
-
-    test("includes iframe-resizer child script in iframe mode", () => {
-      const html = renderTicket(event, { iframe: true });
-      expect(html).toContain("iframe-resizer-child.js");
-    });
-
-    test("excludes iframe-resizer child script when not in iframe mode", () => {
-      const html = renderTicket(event);
-      expect(html).not.toContain("iframe-resizer-child.js");
-    });
-
-    test("renders terms and conditions with checkbox", () => {
-      const html = renderTicket(event, { terms: "No refunds allowed" });
-      expect(html).toContain("No refunds allowed");
-      expect(html).toContain('class="prose"');
-      expect(html).toContain('name="agree_terms"');
-    });
-
-    test("renders markdown paragraphs in terms and conditions", () => {
-      const html = renderTicket(event, {
-        terms: "Line one\n\nLine two\n\nLine three",
-      });
-      expect(html).toContain("<p>Line one</p>");
-      expect(html).toContain("<p>Line two</p>");
-      expect(html).toContain("<p>Line three</p>");
-    });
-
-    test("does not render terms when not provided", () => {
-      const html = renderTicket(event);
-      expect(html).not.toContain('class="terms"');
-      expect(html).not.toContain('name="agree_terms"');
-    });
-
-    test("renders custom questions when provided", () => {
-      const questions = [
-        {
-          id: 1,
-          text: "Size?",
-          answers: [
-            { id: 10, question_id: 1, text: "Small", sort_order: 0 },
-            { id: 11, question_id: 1, text: "Large", sort_order: 1 },
-          ],
-        },
-      ];
-      const html = renderTicket(event, { questions });
-      expect(html).toContain("Size?");
-      expect(html).toContain('name="question_1"');
-    });
-
-    test("includes OpenGraph tags when baseUrl is provided", () => {
-      const ev = testEventWithCount({
-        name: "Birthday Party",
-        slug: "birthday-party",
-        description: "A fun party",
-      });
-      const html = renderTicket(ev, {
-        baseUrl: "https://tix.example.com",
-      });
-      expect(html).toContain(
-        '<meta property="og:title" content="Birthday Party">',
-      );
-      expect(html).toContain('<meta property="og:type" content="website">');
-      expect(html).toContain(
-        '<meta property="og:url" content="https://tix.example.com/ticket/birthday-party">',
-      );
-      expect(html).toContain(
-        '<meta property="og:description" content="A fun party">',
-      );
-    });
-
-    test("does not include OpenGraph tags when baseUrl is not provided", () => {
-      const html = renderTicket(event);
-      expect(html).not.toContain("og:title");
-    });
-  });
-
-  describe("buildOgTags", () => {
-    test("includes title, type, and url", () => {
-      const html = buildOgTags(
-        { name: "My Event", description: "", slug: "my-event", image_url: "" },
-        "https://example.com",
-      );
-      expect(html).toContain('<meta property="og:title" content="My Event">');
-      expect(html).toContain('<meta property="og:type" content="website">');
-      expect(html).toContain(
-        '<meta property="og:url" content="https://example.com/ticket/my-event">',
-      );
-    });
-
-    test("includes description when present", () => {
-      const html = buildOgTags(
-        {
-          name: "My Event",
-          description: "Come join us",
-          slug: "my-event",
-          image_url: "",
-        },
-        "https://example.com",
-      );
-      expect(html).toContain(
-        '<meta property="og:description" content="Come join us">',
-      );
-    });
-
-    test("excludes description when empty", () => {
-      const html = buildOgTags(
-        { name: "My Event", description: "", slug: "my-event", image_url: "" },
-        "https://example.com",
-      );
-      expect(html).not.toContain("og:description");
-    });
-
-    test("includes image when present", () => {
-      const html = buildOgTags(
-        {
-          name: "My Event",
-          description: "",
-          slug: "my-event",
-          image_url: "photo.jpg",
-        },
-        "https://example.com",
-      );
-      expect(html).toContain(
-        '<meta property="og:image" content="https://example.com/image/photo.jpg">',
-      );
-    });
-
-    test("excludes image when empty", () => {
-      const html = buildOgTags(
-        { name: "My Event", description: "", slug: "my-event", image_url: "" },
-        "https://example.com",
-      );
-      expect(html).not.toContain("og:image");
-    });
-
-    test("escapes HTML in event name", () => {
-      const html = buildOgTags(
-        {
-          name: 'Event "with quotes"',
-          description: "",
-          slug: "my-event",
-          image_url: "",
-        },
-        "https://example.com",
-      );
-      expect(html).toContain("Event &quot;with quotes&quot;");
-      expect(html).not.toContain('content="Event "with quotes""');
-    });
-  });
-
-  describe("notFoundPage", () => {
-    test("renders not found message", () => {
-      const html = notFoundPage();
-      expect(html).toContain("<h1>Not Found</h1>");
-    });
-  });
-
-  describe("temporaryErrorPage", () => {
-    test("renders error message with auto-refresh", () => {
-      const html = temporaryErrorPage();
-      expect(html).toContain("<h1>Temporary Error</h1>");
-      expect(html).toContain("Retrying automatically");
-      expect(html).toContain('http-equiv="refresh"');
-      expect(html).toContain('content="2"');
-    });
-  });
-
-  describe("paymentPage", () => {
-    const event = testEvent({ unit_price: 1000 });
-    const attendee = testAttendee();
-
-    test("renders payment details", () => {
-      const html = paymentPage(
-        event,
-        attendee,
-        "https://checkout.stripe.com/session",
-        "£10.00",
-      );
-      expect(html).toContain("Complete Your Payment");
-      expect(html).toContain("John Doe");
-      expect(html).toContain("john@example.com");
-      expect(html).toContain("£10.00");
-    });
-
-    test("includes checkout URL", () => {
-      const html = paymentPage(
-        event,
-        attendee,
-        "https://checkout.stripe.com/session",
-        "£10.00",
-      );
-      expect(html).toContain("https://checkout.stripe.com/session");
-      expect(html).toContain("Pay Now");
-    });
-
-    test("escapes user data", () => {
-      const evilAttendee = testAttendee({ name: "<script>evil()</script>" });
-      const html = paymentPage(
-        event,
-        evilAttendee,
-        "https://checkout.stripe.com/session",
-        "£10.00",
-      );
-      expect(html).toContain("&lt;script&gt;");
-    });
-  });
-
-  describe("successPage", () => {
-    test("renders order success message when paid", () => {
-      const html = successPage({
-        ticketUrl: null,
-        thankYouUrl: "https://example.com/thanks",
-        paid: true,
-      });
-      expect(html).toContain("Thank you for your order");
-      expect(html).toContain("https://example.com/thanks");
-    });
-
-    test("renders order success message when not paid", () => {
-      const html = successPage({ ticketUrl: "/t/abc123" });
-      expect(html).toContain("Order Successful");
-      expect(html).toContain("Thank you for your order");
-    });
-
-    test("includes meta refresh redirect", () => {
-      const html = successPage({
-        ticketUrl: null,
-        thankYouUrl: "https://example.com/thanks",
-        paid: true,
-      });
-      expect(html).toContain('http-equiv="refresh"');
-      expect(html).toContain("3;url=https://example.com/thanks");
-    });
-
-    test("includes data-payment-result attribute for popup postMessage", () => {
-      const html = successPage({ ticketUrl: null, paid: true });
-      expect(html).toContain('data-payment-result="success"');
-    });
-
-    test("excludes data-payment-result attribute when not paid", () => {
-      const html = successPage({ ticketUrl: null });
-      expect(html).not.toContain("data-payment-result");
-    });
-
-    test("renders without redirect when thankYouUrl is empty", () => {
-      const html = successPage({ ticketUrl: null, paid: true });
-      expect(html).not.toContain('http-equiv="refresh"');
-      expect(html).not.toContain("redirected");
-    });
-
-    test("renders ticket link with plural text for multiple tickets", () => {
-      const html = successPage({ ticketUrl: "/t/abc123+def456", paid: true });
-      expect(html).toContain('href="/t/abc123+def456"');
-      expect(html).toContain('target="_blank"');
-      expect(html).toContain("Click here to view your ticket");
-    });
-
-    test("renders ticket link with singular text for single ticket", () => {
-      const html = successPage({ ticketUrl: "/t/abc123", paid: true });
-      expect(html).toContain('href="/t/abc123"');
-      expect(html).toContain("Click here to view your ticket");
-    });
-
-    test("renders both ticket link and redirect when both provided", () => {
-      const html = successPage({
-        ticketUrl: "/t/abc123",
-        thankYouUrl: "https://example.com/thanks",
-        paid: true,
-      });
-      expect(html).toContain('href="/t/abc123"');
-      expect(html).toContain("Click here to view your ticket");
-      expect(html).toContain("https://example.com/thanks");
-      expect(html).toContain('http-equiv="refresh"');
-    });
-
-    test("does not render ticket link when ticketUrl is null", () => {
-      const html = successPage({ ticketUrl: null, paid: true });
-      expect(html).not.toContain("view your ticket");
-    });
-
-    test("includes iframe-resizer child script in iframe mode", () => {
-      detectIframeMode("https://example.com/?iframe=true");
-      const html = successPage({ ticketUrl: "/t/abc123" });
-      expect(html).toContain("iframe-resizer-child.js");
-      expect(html).toContain('class="iframe"');
-      detectIframeMode("https://example.com/");
-    });
-
-    test("excludes iframe-resizer child script when not in iframe mode", () => {
-      const html = successPage({ ticketUrl: "/t/abc123" });
-      expect(html).not.toContain("iframe-resizer-child.js");
-      expect(html).not.toContain('class="iframe"');
-    });
-
-    test("includes scroll-into-view marker in iframe mode", () => {
-      detectIframeMode("https://example.com/?iframe=true");
-      const html = successPage({ ticketUrl: "/t/abc123" });
-      expect(html).toContain("data-scroll-into-view");
-      detectIframeMode("https://example.com/");
-    });
-
-    test("excludes scroll-into-view marker when not in iframe mode", () => {
-      const html = successPage({ ticketUrl: "/t/abc123" });
-      expect(html).not.toContain("data-scroll-into-view");
-    });
-
-    test("shows email notice when fromEmail is provided", () => {
-      const html = successPage({
-        ticketUrl: "/t/abc123",
-        paid: true,
-        fromEmail: "tickets@example.com",
-      });
-      expect(html).toContain("tickets@example.com");
-      expect(html).toContain("Junk/Spam");
-    });
-
-    test("does not show email notice when fromEmail is empty", () => {
-      const html = successPage({ ticketUrl: "/t/abc123", paid: true });
-      expect(html).not.toContain("Junk/Spam");
-    });
-  });
-
-  describe("paymentCancelPage", () => {
-    const event = testEvent({ unit_price: 1000 });
-
-    test("renders cancel message", () => {
-      const html = paymentCancelPage(event, "/ticket/ab12c");
-      expect(html).toContain("Payment Cancelled");
-      expect(html).toContain("/ticket/ab12c");
-      expect(html).toContain("Try again");
-    });
-
-    test("includes data-payment-result attribute for popup postMessage", () => {
-      const html = paymentCancelPage(event, "/ticket/ab12c");
-      expect(html).toContain('data-payment-result="cancel"');
-    });
-  });
-
-  describe("checkoutPopupPage", () => {
-    test("renders checkout URL in data attribute", () => {
-      const html = checkoutPopupPage("https://checkout.stripe.com/session123");
-      expect(html).toContain(
-        'data-checkout-popup="https://checkout.stripe.com/session123"',
-      );
-    });
-
-    test("renders Pay Now link with target _blank", () => {
-      const html = checkoutPopupPage("https://checkout.stripe.com/session123");
-      expect(html).toContain('target="_blank"');
-      expect(html).toContain("Pay Now");
-      expect(html).toContain("data-open-checkout");
-    });
-
-    test("includes waiting element for popup state", () => {
-      const html = checkoutPopupPage("https://checkout.stripe.com/session123");
-      expect(html).toContain("data-checkout-waiting");
-    });
-
-    test("uses iframe body class", () => {
-      const html = checkoutPopupPage("https://checkout.stripe.com/session123");
-      expect(html).toContain('class="iframe"');
-    });
-
-    test("includes iframe-resizer child script", () => {
-      const html = checkoutPopupPage("https://checkout.stripe.com/session123");
-      expect(html).toContain("iframe-resizer-child.js");
-    });
-
-    test("escapes checkout URL", () => {
-      const html = checkoutPopupPage('https://evil.com/"onload="alert(1)');
-      expect(html).toContain("&quot;");
-      expect(html).not.toContain('"onload="');
-    });
-
-    test("includes scroll-into-view marker for parent scroll", () => {
-      const html = checkoutPopupPage("https://checkout.stripe.com/session123");
-      expect(html).toContain("data-scroll-into-view");
-    });
-  });
-
-  test("shows email notice for reservation when fromEmail is provided", () => {
-    const html = successPage({
-      ticketUrl: "/t/abc123",
-      fromEmail: "tickets@example.com",
-    });
-    expect(html).toContain("tickets@example.com");
-    expect(html).toContain("Junk/Spam");
-  });
-
-  test("does not show email notice for reservation when fromEmail is empty", () => {
-    const html = successPage({ ticketUrl: "/t/abc123" });
-    expect(html).not.toContain("Junk/Spam");
-  });
-
-  describe("paymentErrorPage", () => {
-    test("renders error message", () => {
-      const html = paymentErrorPage("Payment verification failed");
-      expect(html).toContain("Payment Error");
-      expect(html).toContain("Payment verification failed");
-      expect(html).toContain('class="error"');
-    });
-
-    test("escapes error message", () => {
-      const html = paymentErrorPage("<script>evil()</script>");
-      expect(html).toContain("&lt;script&gt;");
-    });
-
-    test("includes home link", () => {
-      const html = paymentErrorPage("Error");
-      expect(html).toContain('href="/"');
-    });
-  });
-
-  describe("adminEventNewPage", () => {
-    test("renders create event form fields", () => {
-      const html = adminEventNewPage([], TEST_SESSION);
-      expect(html).toContain("Add Event");
-      expect(html).toContain('name="name"');
-      expect(html).toContain('name="max_attendees"');
-      expect(html).toContain('name="thank_you_url"');
-      expect(html).toContain('name="unit_price"');
-      expect(html).toContain("Ticket Price");
-    });
-
-    test("renders breadcrumb back link", () => {
-      const html = adminEventNewPage([], TEST_SESSION);
-      expect(html).toContain('href="/admin/"');
-      expect(html).toContain("Events");
-    });
-
-    test("renders group select when groups exist", () => {
-      const groups = [testGroup({ id: 2, name: "My Group" })];
-      const html = adminEventNewPage(groups, TEST_SESSION);
-      expect(html).toContain('name="group_id"');
-      expect(html).toContain('<option value="0" selected>No Group</option>');
-      expect(html).toContain('<option value="2"');
-      expect(html).toContain("My Group");
-    });
-
-    test("renders error when provided", () => {
-      const html = adminEventNewPage([], TEST_SESSION, "Something went wrong");
-      expect(html).toContain("Something went wrong");
-    });
-  });
-
-  describe("adminEventPage export button", () => {
-    test("renders export CSV button", () => {
-      const event = testEventWithCount({ attendee_count: 2 });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("/admin/event/1/export");
-      expect(html).toContain("Export CSV");
-    });
-  });
-
-  describe("generateAttendeesCsv", () => {
-    test("generates CSV header for empty attendees", () => {
-      const csv = generateAttendeesCsv([]);
-      expect(csv).toBe(
-        "Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
-      );
-    });
-
-    test("generates CSV with attendee data", () => {
-      const attendees = [
-        testAttendee({ created: "2024-01-15T10:30:00Z", quantity: 2 }),
-      ];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[0]).toBe(
-        "Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
-      );
-      expect(lines[1]).toContain("John Doe");
-      expect(lines[1]).toContain("john@example.com");
-      expect(lines[1]).toContain(",2,");
-      expect(lines[1]).toContain("2024-01-15T10:30:00.000Z");
-    });
-
-    test("escapes values with commas", () => {
-      const attendees = [testAttendee({ name: "Doe, John" })];
-      const csv = generateAttendeesCsv(attendees);
-      expect(csv).toContain('"Doe, John"');
-    });
-
-    test("escapes values with quotes", () => {
-      const attendees = [testAttendee({ name: 'John "JD" Doe' })];
-      const csv = generateAttendeesCsv(attendees);
-      expect(csv).toContain('"John ""JD"" Doe"');
-    });
-
-    test("escapes values with newlines", () => {
-      const attendees = [testAttendee({ name: "John\nDoe" })];
-      const csv = generateAttendeesCsv(attendees);
-      expect(csv).toContain('"John\nDoe"');
-    });
-
-    test("generates multiple rows", () => {
-      const attendees = [
-        testAttendee(),
-        testAttendee({
-          id: 2,
-          name: "Jane Smith",
-          email: "jane@example.com",
-          created: "2024-01-16T11:00:00Z",
-          quantity: 3,
-        }),
-      ];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines).toHaveLength(3);
-      expect(lines[1]).toContain("John Doe");
-      expect(lines[2]).toContain("Jane Smith");
-    });
-
-    test("includes phone number in CSV output", () => {
-      const attendees = [testAttendee({ phone: "+1 555 123 4567" })];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[0]).toBe(
-        "Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
-      );
-      expect(lines[1]).toContain("+1 555 123 4567");
-    });
-
-    test("includes empty phone column when phone not collected", () => {
-      const attendees = [testAttendee()];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain("john@example.com,,,,1,");
-    });
-
-    test("generates CSV with price and transaction ID", () => {
-      const attendees = [
-        testAttendee({
-          payment_id: "pi_abc123",
-          quantity: 2,
-          price_paid: "2000",
-        }),
-      ];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain("20.00");
-      expect(lines[1]).toContain("pi_abc123");
-    });
-
-    test("formats zero price and empty payment_id for free attendees", () => {
-      const attendees = [testAttendee()];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      // Price should be 0.00, transaction ID empty, followed by Checked In, Token, URL
-      expect(lines[1]).toContain("0.00,,No,");
-    });
-
-    test("shared transaction ID across multiple attendees", () => {
-      const attendees = [
-        testAttendee({
-          payment_id: "pi_shared_123",
-          price_paid: "1000",
-        }),
-        testAttendee({
-          id: 2,
-          event_id: 2,
-          payment_id: "pi_shared_123",
-          quantity: 2,
-          price_paid: "3000",
-        }),
-      ];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain("10.00");
-      expect(lines[1]).toContain("pi_shared_123");
-      expect(lines[2]).toContain("30.00");
-      expect(lines[2]).toContain("pi_shared_123");
-    });
-
-    test("includes Checked In as Yes for checked-in attendee", () => {
-      const attendees = [testAttendee({ checked_in: true })];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain(",Yes,");
-    });
-
-    test("includes Checked In as No for not checked-in attendee", () => {
-      const attendees = [testAttendee({ checked_in: false })];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain(",No,");
-    });
-
-    test("includes ticket token and URL in CSV output", () => {
-      const attendees = [testAttendee({ ticket_token: "abc123" })];
-      const csv = generateAttendeesCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain("abc123");
-      expect(lines[1]).toContain("https://localhost/t/abc123");
-    });
-
-    test("includes Date column when includeDate is true", () => {
-      const csv = generateAttendeesCsv([], true);
-      expect(csv).toBe(
-        "Date,Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
-      );
-    });
-
-    test("includes date value in row when includeDate is true", () => {
-      const attendees = [testAttendee({ date: "2026-03-15" })];
-      const csv = generateAttendeesCsv(attendees, true);
-      const lines = csv.split("\n");
-      expect(lines[0]).toContain("Date,Name");
-      expect(lines[1]).toMatch(/^2026-03-15,/);
-    });
-
-    test("includes empty date in row when date is null", () => {
-      const attendees = [testAttendee({ date: null })];
-      const csv = generateAttendeesCsv(attendees, true);
-      const lines = csv.split("\n");
-      expect(lines[1]).toMatch(/^,John Doe/);
-    });
-
-    test("omits Date column when includeDate is false", () => {
-      const attendees = [testAttendee({ date: "2026-03-15" })];
-      const csv = generateAttendeesCsv(attendees, false);
-      expect(csv.startsWith("Name,")).toBe(true);
-      expect(csv).not.toContain("2026-03-15");
-    });
-  });
-
-  describe("adminEventPage filter links", () => {
-    test("renders All / Checked In / Checked Out links", () => {
-      const event = testEventWithCount({ attendee_count: 1 });
-      const attendees = [testAttendee()];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("All");
-      expect(html).toContain("Checked In");
-      expect(html).toContain("Checked Out");
-    });
-
-    test("bolds All when no filter is active", () => {
-      const event = testEventWithCount({ attendee_count: 1 });
-      const attendees = [testAttendee()];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("<strong>All</strong>");
-      expect(html).toContain(`href="/admin/event/${event.id}/in#attendees"`);
-      expect(html).toContain(`href="/admin/event/${event.id}/out#attendees"`);
-    });
-
-    test("bolds Checked In when filter is in", () => {
-      const event = testEventWithCount({ attendee_count: 1 });
-      const attendees = [testAttendee()];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        activeFilter: "in",
-      });
-      expect(html).toContain("<strong>Checked In</strong>");
-      expect(html).toContain(`href="/admin/event/${event.id}#attendees"`);
-    });
-
-    test("bolds Checked Out when filter is out", () => {
-      const event = testEventWithCount({ attendee_count: 1 });
-      const attendees = [testAttendee()];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        activeFilter: "out",
-      });
-      expect(html).toContain("<strong>Checked Out</strong>");
-    });
-
-    test("filters to only checked-in attendees when filter is in", () => {
-      const event = testEventWithCount({ attendee_count: 2 });
-      const attendees = [
-        testAttendee({ id: 1, name: "Checked In User", checked_in: true }),
-        testAttendee({ id: 2, name: "Not Checked In User", checked_in: false }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        activeFilter: "in",
-      });
-      expect(html).toContain("Checked In User");
-      expect(html).not.toContain("Not Checked In User");
-    });
-
-    test("filters to only checked-out attendees when filter is out", () => {
-      const event = testEventWithCount({ attendee_count: 2 });
-      const attendees = [
-        testAttendee({ id: 1, name: "Alice InPerson", checked_in: true }),
-        testAttendee({ id: 2, name: "Bob Remote", checked_in: false }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        activeFilter: "out",
-      });
-      expect(html).not.toContain("Alice InPerson");
-      expect(html).toContain("Bob Remote");
-    });
-
-    test("shows all attendees when filter is all", () => {
-      const event = testEventWithCount({ attendee_count: 2 });
-      const attendees = [
-        testAttendee({ id: 1, name: "Checked In User", checked_in: true }),
-        testAttendee({ id: 2, name: "Not Checked In User", checked_in: false }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        activeFilter: "all",
-      });
-      expect(html).toContain("Checked In User");
-      expect(html).toContain("Not Checked In User");
-    });
-
-    test("includes return_filter hidden field in checkin form", () => {
-      const event = testEventWithCount({ attendee_count: 1 });
-      const attendees = [testAttendee({ checked_in: true })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-        activeFilter: "in",
-      });
-      expect(html).toContain('name="return_filter"');
-      expect(html).toContain('value="in"');
-    });
-  });
-
-  describe("adminEventActivityLogPage", () => {
-    test("renders activity log entries", () => {
-      const event = testEventWithCount();
-      const entries = [
-        {
-          id: 1,
-          created: "2024-01-15T10:30:00Z",
-          event_id: 1,
-          message: "Ticket reserved",
-        },
-        {
-          id: 2,
-          created: "2024-01-15T11:00:00Z",
-          event_id: 1,
-          message: "Payment received",
-        },
-      ];
-      const html = adminEventActivityLogPage(event, entries, TEST_SESSION);
-      expect(html).toContain("Ticket reserved");
-      expect(html).toContain("Payment received");
-      expect(html).toContain("Log");
-    });
-
-    test("renders empty state when no entries", () => {
-      const event = testEventWithCount();
-      const html = adminEventActivityLogPage(event, [], TEST_SESSION);
-      expect(html).toContain("No activity recorded yet");
-    });
-  });
-
-  describe("adminGlobalActivityLogPage", () => {
-    test("renders global activity log with entries", () => {
-      const entries = [
-        {
-          id: 1,
-          created: "2024-01-15T10:30:00Z",
-          event_id: null,
-          message: "System started",
-        },
-      ];
-      const html = adminGlobalActivityLogPage(entries, false, TEST_SESSION);
-      expect(html).toContain("System started");
-      expect(html).toContain("Log");
-    });
-
-    test("renders empty state when no entries", () => {
-      const html = adminGlobalActivityLogPage([], false, TEST_SESSION);
-      expect(html).toContain("No activity recorded yet");
-    });
-
-    test("shows truncation message when truncated", () => {
-      const entries = [
-        {
-          id: 1,
-          created: "2024-01-15T10:30:00Z",
-          event_id: null,
-          message: "Action",
-        },
-      ];
-      const html = adminGlobalActivityLogPage(entries, true, TEST_SESSION);
-      expect(html).toContain("Showing the most recent 200 entries");
-    });
-
-    test("does not show truncation message when not truncated", () => {
-      const entries = [
-        {
-          id: 1,
-          created: "2024-01-15T10:30:00Z",
-          event_id: null,
-          message: "Action",
-        },
-      ];
-      const html = adminGlobalActivityLogPage(entries, false, TEST_SESSION);
-      expect(html).not.toContain("Showing the most recent 200 entries");
-    });
-  });
-
-  describe("adminDashboardPage inactive events", () => {
-    test("renders inactive event with reduced opacity", () => {
-      const events = [testEventWithCount({ active: false, attendee_count: 5 })];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("inactive-row");
-      expect(html).toContain("Inactive");
-    });
-  });
-
-  describe("activeEventStatsSection", () => {
-    test("renders income, tickets, and attendees", () => {
-      const html = activeEventStatsSection({
-        income: 5000,
-        tickets: 30,
-        attendees: 52,
-      });
-      expect(html).toContain("Active Event Statistics");
-      expect(html).toContain("<strong>Income:</strong>");
-      expect(html).toContain("<strong>Tickets:</strong>");
-      expect(html).toContain("<strong>Attendees:</strong>");
-      expect(html).toContain("30");
-      expect(html).toContain("52");
-    });
-
-    test("renders zero values", () => {
-      const html = activeEventStatsSection({
-        income: 0,
-        tickets: 0,
-        attendees: 0,
-      });
-      expect(html).toContain("<strong>Tickets:</strong> 0");
-      expect(html).toContain("<strong>Attendees:</strong> 0");
-    });
-
-    test("renders as closed details element", () => {
-      const html = activeEventStatsSection({
-        income: 0,
-        tickets: 0,
-        attendees: 0,
-      });
-      expect(html).toContain("<details>");
-      expect(html).not.toContain("<details open");
-    });
-  });
-
-  describe("adminDashboardPage active event statistics", () => {
-    test("shows stats section when stats provided", () => {
-      const html = adminDashboardPage(
-        [],
-        TEST_SESSION,
-        undefined,
-        [],
-        undefined,
-        {
-          income: 1000,
-          tickets: 5,
-          attendees: 10,
-        },
-      );
-      expect(html).toContain("Active Event Statistics");
-    });
-
-    test("does not show stats section when stats is null", () => {
-      const html = adminDashboardPage(
-        [],
-        TEST_SESSION,
-        undefined,
-        [],
-        undefined,
-        null,
-      );
-      expect(html).not.toContain("Active Event Statistics");
-    });
-
-    test("does not show stats section when stats not provided", () => {
-      const html = adminDashboardPage([], TEST_SESSION);
-      expect(html).not.toContain("Active Event Statistics");
-    });
-  });
-
-  describe("adminDashboardPage multi-booking link", () => {
-    test("does not show multi-booking section with zero events", () => {
-      const html = adminDashboardPage([], TEST_SESSION);
-      expect(html).not.toContain("Multi-booking link");
-    });
-
-    test("does not show multi-booking section with one active event", () => {
-      const events = [testEventWithCount({ id: 1, slug: "ab12c" })];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).not.toContain("Multi-booking link");
-    });
-
-    test("shows multi-booking section with two active events", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c", name: "Event A" }),
-        testEventWithCount({ id: 2, slug: "cd34e", name: "Event B" }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("Multi-booking link");
-      expect(html).toContain("Event A");
-      expect(html).toContain("Event B");
-    });
-
-    test("does not count inactive events toward threshold", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c", active: true }),
-        testEventWithCount({ id: 2, slug: "cd34e", active: false }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).not.toContain("Multi-booking link");
-    });
-
-    test("excludes inactive events from checkboxes", () => {
-      const events = [
-        testEventWithCount({
-          id: 1,
-          slug: "ab12c",
-          name: "Active One",
-          active: true,
-        }),
-        testEventWithCount({
-          id: 2,
-          slug: "cd34e",
-          name: "Inactive",
-          active: false,
-        }),
-        testEventWithCount({
-          id: 3,
-          slug: "ef56g",
-          name: "Active Two",
-          active: true,
-        }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("Active One");
-      expect(html).toContain("Active Two");
-      expect(html).not.toContain('data-multi-booking-slug="cd34e"');
-    });
-
-    test("renders checkboxes with slug data attributes", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c" }),
-        testEventWithCount({ id: 2, slug: "cd34e" }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain('data-multi-booking-slug="ab12c"');
-      expect(html).toContain('data-multi-booking-slug="cd34e"');
-    });
-
-    test("renders URL input with domain data attribute", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c" }),
-        testEventWithCount({ id: 2, slug: "cd34e" }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain('data-domain="localhost"');
-      expect(html).toContain("data-multi-booking-url");
-      expect(html).toContain("readonly");
-      expect(html).toContain('for="multi-booking-url"');
-      expect(html).toContain('id="multi-booking-url"');
-    });
-
-    test("is collapsed by default via details element", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c" }),
-        testEventWithCount({ id: 2, slug: "cd34e" }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("<details>");
-      expect(html).toContain("<summary>");
-    });
-
-    test("renders embed code inputs", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c", fields: "email" }),
-        testEventWithCount({ id: 2, slug: "cd34e", fields: "email,phone" }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain("data-multi-booking-embed-script");
-      expect(html).toContain("data-multi-booking-embed-iframe");
-      expect(html).toContain('for="multi-booking-embed-script"');
-      expect(html).toContain('for="multi-booking-embed-iframe"');
-      expect(html).toContain('id="multi-booking-embed-script"');
-      expect(html).toContain('id="multi-booking-embed-iframe"');
-    });
-
-    test("checkboxes include data-fields attribute for embed code generation", () => {
-      const events = [
-        testEventWithCount({ id: 1, slug: "ab12c", fields: "email" }),
-        testEventWithCount({ id: 2, slug: "cd34e", fields: "email,phone" }),
-      ];
-      const html = adminDashboardPage(events, TEST_SESSION);
-      expect(html).toContain('data-fields="email"');
-      expect(html).toContain('data-fields="email,phone"');
-    });
-  });
-
-  describe("calculateTotalRevenue", () => {
-    test("returns 0 for empty attendees", () => {
-      expect(calculateTotalRevenue([])).toBe(0);
-    });
-
-    test("sums price_paid from attendees", () => {
-      const attendees = [
-        testAttendee({ price_paid: "1000" }),
-        testAttendee({ id: 2, price_paid: "2000" }),
-      ];
-      expect(calculateTotalRevenue(attendees)).toBe(3000);
-    });
-
-    test("returns 0 when attendees have no price_paid", () => {
-      const attendees = [testAttendee({ quantity: 3 })];
-      expect(calculateTotalRevenue(attendees)).toBe(0);
-    });
-
-    test("skips attendees without price_paid when summing", () => {
-      const attendees = [
-        testAttendee({ price_paid: "1500" }),
-        testAttendee({ id: 2, quantity: 2 }),
-      ];
-      expect(calculateTotalRevenue(attendees)).toBe(1500);
-    });
-  });
-
-  describe("countCheckedIn", () => {
-    test("returns 0 for empty attendees", () => {
-      expect(countCheckedIn([])).toBe(0);
-    });
-
-    test("sums quantity for attendees with checked_in true", () => {
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true }),
-        testAttendee({ id: 2, checked_in: false }),
-        testAttendee({ id: 3, checked_in: true }),
-      ];
-      expect(countCheckedIn(attendees)).toBe(2);
-    });
-
-    test("returns 0 when none checked in", () => {
-      const attendees = [
-        testAttendee({ id: 1, checked_in: false }),
-        testAttendee({ id: 2, checked_in: false }),
-      ];
-      expect(countCheckedIn(attendees)).toBe(0);
-    });
-
-    test("sums quantity for multi-quantity checked-in registrations", () => {
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true, quantity: 3 }),
-        testAttendee({ id: 2, checked_in: false, quantity: 2 }),
-        testAttendee({ id: 3, checked_in: true, quantity: 1 }),
-      ];
-      expect(countCheckedIn(attendees)).toBe(4);
-    });
-  });
-
-  describe("countCheckedInRows", () => {
-    test("returns 0 for empty attendees", () => {
-      expect(countCheckedInRows([])).toBe(0);
-    });
-
-    test("counts rows regardless of quantity", () => {
-      const attendees = [
-        testAttendee({ id: 1, checked_in: true, quantity: 3 }),
-        testAttendee({ id: 2, checked_in: false, quantity: 2 }),
-        testAttendee({ id: 3, checked_in: true, quantity: 1 }),
-      ];
-      expect(countCheckedInRows(attendees)).toBe(2);
-    });
-  });
-
-  describe("adminEventPage total revenue", () => {
-    test("shows total revenue for paid events", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 2 });
-      const attendees = [
-        testAttendee({ price_paid: "1000", payment_id: "pi_test_1" }),
-        testAttendee({ id: 2, price_paid: "2000", payment_id: "pi_test_2" }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Total Revenue");
-      expect(html).toContain("£30");
-    });
-
-    test("does not show total revenue for free events", () => {
-      const event = testEventWithCount({ unit_price: 0, attendee_count: 1 });
-      const attendees = [testAttendee()];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("Total Revenue");
-    });
-
-    test("shows zero revenue for paid event with no attendees", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 0 });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Total Revenue");
-      expect(html).toContain("£0");
-    });
-  });
-
-  describe("adminEventPage optional fields", () => {
-    test("shows reactivate link for inactive events", () => {
-      const event = testEventWithCount({ active: false, attendee_count: 0 });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("/reactivate");
-      expect(html).toContain("Reactivate");
-    });
-
-    test("hides thank you URL row when no thank_you_url", () => {
-      const event = testEventWithCount({
-        thank_you_url: "",
-        attendee_count: 0,
-      });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("Thank You URL");
-    });
-
-    test("shows webhook URL in copyable input when present", () => {
-      const event = testEventWithCount({
-        webhook_url: "https://hooks.example.com/notify",
-        attendee_count: 0,
-      });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Webhook URL");
-      expect(html).toContain('value="https://hooks.example.com/notify"');
-      expect(html).toContain("readonly");
-    });
-  });
-
-  describe("nearCapacity", () => {
-    test("returns true when at 90% capacity", () => {
-      const event = testEventWithCount({
-        attendee_count: 90,
-        max_attendees: 100,
-      });
-      expect(nearCapacity(event)).toBe(true);
-    });
-
-    test("returns true when over 90% capacity", () => {
-      const event = testEventWithCount({
-        attendee_count: 95,
-        max_attendees: 100,
-      });
-      expect(nearCapacity(event)).toBe(true);
-    });
-
-    test("returns false when under 90% capacity", () => {
-      const event = testEventWithCount({
-        attendee_count: 89,
-        max_attendees: 100,
-      });
-      expect(nearCapacity(event)).toBe(false);
-    });
-
-    test("returns true when fully sold out", () => {
-      const event = testEventWithCount({
-        attendee_count: 100,
-        max_attendees: 100,
-      });
-      expect(nearCapacity(event)).toBe(true);
-    });
-  });
-
-  describe("isIncompletePayment", () => {
-    test("returns true for paid event attendee with no payment_id and price > 0", () => {
-      const attendee = testAttendee({ payment_id: "", price_paid: "1000" });
-      expect(isIncompletePayment(attendee, true)).toBe(true);
-    });
-
-    test("returns false for free event", () => {
-      const attendee = testAttendee({ payment_id: "", price_paid: "0" });
-      expect(isIncompletePayment(attendee, false)).toBe(false);
-    });
-
-    test("returns false for admin-added attendee on paid event (price_paid=0)", () => {
-      const attendee = testAttendee({ payment_id: "", price_paid: "0" });
-      expect(isIncompletePayment(attendee, true)).toBe(false);
-    });
-
-    test("returns false for completed payment attendee", () => {
-      const attendee = testAttendee({
-        payment_id: "pi_test_123",
-        price_paid: "1000",
-      });
-      expect(isIncompletePayment(attendee, true)).toBe(false);
-    });
-  });
-
-  describe("adminEventPage failed payments", () => {
-    test("shows Failed Payments section when incomplete attendees exist", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 3 });
-      const attendees = [
-        testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-        testAttendee({ id: 2, payment_id: "", price_paid: "1000" }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Failed Payments");
-      expect(html).toContain("1 attendee(s) with unresolved payments");
-      expect(html).toContain("/delete-incomplete");
-    });
-
-    test("hides Failed Payments section when no incomplete attendees", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 1 });
-      const attendees = [
-        testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("Failed Payments");
-    });
-
-    test("hides Failed Payments section for free events", () => {
-      const event = testEventWithCount({ unit_price: 0, attendee_count: 1 });
-      const attendees = [testAttendee({ id: 1, price_paid: "0" })];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("Failed Payments");
-    });
-
-    test("excludes incomplete attendees from attendee count", () => {
-      const event = testEventWithCount({
-        unit_price: 1000,
-        attendee_count: 3,
-        max_attendees: 100,
-      });
-      const attendees = [
-        testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-        testAttendee({ id: 2, payment_id: "", price_paid: "1000" }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      // adjusted count: 3 - 1 (incomplete qty) = 2
-      expect(html).toContain("2 / 100");
-    });
-
-    test("excludes incomplete attendees from checked-in count", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 2 });
-      const attendees = [
-        testAttendee({
-          id: 1,
-          payment_id: "pi_ok",
-          price_paid: "1000",
-          checked_in: true,
-        }),
-        testAttendee({
-          id: 2,
-          payment_id: "",
-          price_paid: "1000",
-          checked_in: true,
-        }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      // Only complete attendees count: 1 checked in / 1 total
-      expect(html).toContain("1 / 1");
-    });
-
-    test("excludes incomplete attendees from revenue", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 2 });
-      const attendees = [
-        testAttendee({ id: 1, payment_id: "pi_ok", price_paid: "1000" }),
-        testAttendee({ id: 2, payment_id: "", price_paid: "2000" }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("£10");
-      expect(html).not.toContain("£30");
-    });
-
-    test("failed payments table has delete button but no check-in or refund", () => {
-      const event = testEventWithCount({ unit_price: 1000, attendee_count: 1 });
-      const attendees = [
-        testAttendee({
-          id: 1,
-          name: "Jane Stuck",
-          payment_id: "",
-          price_paid: "1000",
-        }),
-      ];
-      const html = adminEventPage({
-        event,
-        attendees,
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      const failedSection =
-        html.split("Failed Payments")[1]?.split("Add Attendee")[0] ?? "";
-      expect(failedSection).toContain("Jane Stuck");
-      expect(failedSection).toContain("Delete");
-      expect(failedSection).toContain("/delete-incomplete");
-      expect(failedSection).not.toContain("Check in");
-      expect(failedSection).not.toContain("Check out");
-      expect(failedSection).not.toContain("Refund");
-      expect(failedSection).not.toContain("Re-send Webhook");
-    });
-  });
-
-  describe("formatAddressInline", () => {
-    test("returns empty string for empty input", () => {
-      expect(formatAddressInline("")).toBe("");
-    });
-
-    test("returns single line unchanged", () => {
-      expect(formatAddressInline("123 Main St")).toBe("123 Main St");
-    });
-
-    test("joins multiple lines with comma-space", () => {
-      expect(formatAddressInline("123 Main St\nSpringfield\nIL 62701")).toBe(
-        "123 Main St, Springfield, IL 62701",
-      );
-    });
-
-    test("handles Windows line endings (CRLF)", () => {
-      expect(formatAddressInline("123 Main St\r\nSpringfield")).toBe(
-        "123 Main St, Springfield",
-      );
-    });
-
-    test("does not double-comma when line ends with comma", () => {
-      expect(formatAddressInline("123 Main St,\nSpringfield")).toBe(
-        "123 Main St, Springfield",
-      );
-    });
-
-    test("trims whitespace from lines", () => {
-      expect(formatAddressInline("  123 Main St  \n  Springfield  ")).toBe(
-        "123 Main St, Springfield",
-      );
-    });
-
-    test("skips blank lines", () => {
-      expect(formatAddressInline("123 Main St\n\n\nSpringfield")).toBe(
-        "123 Main St, Springfield",
-      );
-    });
-  });
-
-  describe("adminSessionsPage", () => {
-    test("renders session rows", () => {
-      const sessions = [
-        {
-          token: "abcdefghijklmnop",
-          csrf_token: "csrf1",
-          expires: Date.now() + 86400000,
-          wrapped_data_key: null,
-          user_id: 1,
-        },
-        {
-          token: "qrstuvwxyz123456",
-          csrf_token: "csrf2",
-          expires: Date.now() + 86400000,
-          wrapped_data_key: null,
-          user_id: 2,
-        },
-      ];
-      const html = adminSessionsPage(
-        sessions,
-        "abcdefghijklmnop",
-        TEST_SESSION,
-        "",
-      );
-      expect(html).toContain("abcdefgh...");
-      expect(html).toContain("qrstuvwx...");
-      expect(html).toContain("Current");
-    });
-
-    test("renders empty state when no sessions", () => {
-      const html = adminSessionsPage([], "some-token", TEST_SESSION, "");
-      expect(html).toContain("No sessions");
-    });
-  });
-
-  describe("adminUsersPage", () => {
-    test("renders statuses and actions for different user states", () => {
-      const users: DisplayUser[] = [
-        {
-          id: 1,
-          username: "owner",
-          adminLevel: "owner",
-          hasPassword: true,
-          hasDataKey: true,
-          inviteExpired: false,
-        },
-        {
-          id: 2,
-          username: "pending",
-          adminLevel: "manager",
-          hasPassword: true,
-          hasDataKey: false,
-          inviteExpired: false,
-        },
-        {
-          id: 3,
-          username: "invited",
-          adminLevel: "manager",
-          hasPassword: false,
-          hasDataKey: false,
-          inviteExpired: false,
-        },
-      ];
-      const html = adminUsersPage(users, TEST_SESSION, {
-        inviteLink: "",
-        success: "",
-        error: "",
-        currentUserId: 1,
-      });
-      expect(html).toContain("Active");
-      expect(html).toContain("Pending Activation");
-      expect(html).toContain("Invited");
-      expect(html).toContain('action="/admin/users/2/activate"');
-      expect(html).toContain('href="/admin/users/2/delete"');
-      expect(html).toContain('href="/admin/users/3/delete"');
-      expect(html).not.toContain('href="/admin/users/1/delete"');
-    });
-
-    test("renders Invite Expired status for expired invite", () => {
-      const users: DisplayUser[] = [
-        {
-          id: 1,
-          username: "owner",
-          adminLevel: "owner",
-          hasPassword: true,
-          hasDataKey: true,
-          inviteExpired: false,
-        },
-        {
-          id: 2,
-          username: "expired-user",
-          adminLevel: "manager",
-          hasPassword: false,
-          hasDataKey: false,
-          inviteExpired: true,
-        },
-      ];
-      const html = adminUsersPage(users, TEST_SESSION, {
-        inviteLink: "",
-        success: "",
-        error: "",
-        currentUserId: 1,
-      });
-      expect(html).toContain("Invite Expired");
-    });
-
-    test("renders invite, success, and error messages when provided", () => {
-      const users: DisplayUser[] = [
-        {
-          id: 1,
-          username: "owner",
-          adminLevel: "owner",
-          hasPassword: true,
-          hasDataKey: true,
-          inviteExpired: false,
-        },
-      ];
-      const html = adminUsersPage(users, TEST_SESSION, {
-        inviteLink: "https://example.com/join/abc123",
-        success: "Invite created",
-        error: "Something went wrong",
-        currentUserId: 1,
-      });
-      expect(html).toContain("Invite link (share this with the new user)");
-      expect(html).toContain("https://example.com/join/abc123");
-      expect(html).toContain("Invite created");
-      expect(html).toContain("Something went wrong");
-    });
-  });
-
-  describe("adminUserDeletePage", () => {
-    test("renders delete confirmation form with username", () => {
-      const user: DisplayUser = {
-        id: 5,
-        username: "targetuser",
-        adminLevel: "manager",
-        hasPassword: true,
-        hasDataKey: true,
-        inviteExpired: false,
-      };
-      const html = adminUserDeletePage(user, TEST_SESSION);
-      expect(html).toContain("Delete User");
-      expect(html).toContain("targetuser");
-      expect(html).toContain('name="confirm_identifier"');
-      expect(html).toContain('action="/admin/users/5/delete"');
-      expect(html).toContain("permanently delete");
-    });
-
-    test("renders error message when provided", () => {
-      const user: DisplayUser = {
-        id: 5,
-        username: "targetuser",
-        adminLevel: "owner",
-        hasPassword: true,
-        hasDataKey: true,
-        inviteExpired: false,
-      };
-      const html = adminUserDeletePage(
-        user,
-        TEST_SESSION,
-        "Username does not match",
-      );
-      expect(html).toContain("Username does not match");
-    });
-  });
-
-  describe("ticketPage", () => {
-    test("shows all sold out message when every event is sold out", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 100,
-            max_attendees: 100,
-          }),
-        ),
-        buildTicketEvent(
-          testEventWithCount({
-            id: 2,
-            slug: "cd34e",
-            name: "Event B",
-            attendee_count: 50,
-            max_attendees: 50,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c", "cd34e"] });
-      expect(html).toContain("Sorry, all events are sold out.");
-      expect(html).not.toContain("Reserve Tickets</button>");
-    });
-
-    test("renders markdown paragraphs in terms and conditions", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-          }),
-        ),
-      ];
-      const html = ticketPage({
-        events,
-        slugs: ["ab12c"],
-        terms: "Rule one\n\nRule two",
-      });
-      expect(html).toContain("<p>Rule one</p>");
-      expect(html).toContain("<p>Rule two</p>");
-      expect(html).toContain('name="agree_terms"');
-    });
-
-    test("renders custom questions with event IDs", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-          }),
-        ),
-      ];
-      const questions = [
-        {
-          id: 5,
-          text: "Size?",
-          answers: [{ id: 10, question_id: 5, text: "Small", sort_order: 0 }],
-        },
-      ];
-      const questionEventMap = new Map([[5, [1]]]);
-      const html = ticketPage({
-        events,
-        slugs: ["ab12c"],
-        questions,
-        questionEventMap,
-      });
-      expect(html).toContain("Size?");
-      expect(html).toContain('name="question_5"');
-      expect(html).toContain('data-event-ids="1"');
-    });
-
-    test("appends ?iframe=true to form action in iframe mode", () => {
-      detectIframeMode("https://example.com/?iframe=true");
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c"] });
-      expect(html).toContain('action="/ticket/ab12c?iframe=true"');
-      expect(html).toContain('class="iframe"');
-      detectIframeMode("https://example.com/");
-    });
-
-    test("includes iframe-resizer child script in iframe mode", () => {
-      detectIframeMode("https://example.com/?iframe=true");
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c"] });
-      expect(html).toContain("iframe-resizer-child.js");
-      detectIframeMode("https://example.com/");
-    });
-
-    test("excludes iframe-resizer child script without iframe mode", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c"] });
-      expect(html).not.toContain("iframe-resizer-child.js");
-    });
-
-    test("does not append ?iframe=true without iframe mode", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c"] });
-      expect(html).toContain('action="/ticket/ab12c"');
-      expect(html).not.toContain("?iframe=true");
-      expect(html).not.toContain('class="iframe"');
-    });
-
-    test("hides quantity selector for single event with max quantity 1", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-            max_quantity: 1,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c"] });
-      expect(html).toContain('name="quantity_1" value="1"');
-      expect(html).not.toContain("<select");
-      expect(html).not.toContain("Select Tickets");
-    });
-
-    test("shows quantity selector for single event with max quantity above 1", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-            max_quantity: 3,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c"] });
-      expect(html).toContain("<select");
-      expect(html).toContain('name="quantity_1"');
-      expect(html).toContain("Number of Tickets");
-      expect(html).not.toContain('name="quantity_1" value="1"');
-    });
-
-    test("shows quantity selector for multiple events even with max quantity 1", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-            max_quantity: 1,
-          }),
-        ),
-        buildTicketEvent(
-          testEventWithCount({
-            id: 2,
-            slug: "cd34e",
-            name: "Event B",
-            attendee_count: 0,
-            max_quantity: 1,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c", "cd34e"] });
-      expect(html).toContain("<select");
-      expect(html).toContain("Select Tickets");
-    });
-
-    test("hides quantity selector when one event available and one sold out", () => {
-      const events = [
-        buildTicketEvent(
-          testEventWithCount({
-            id: 1,
-            slug: "ab12c",
-            name: "Event A",
-            attendee_count: 0,
-            max_quantity: 1,
-          }),
-        ),
-        buildTicketEvent(
-          testEventWithCount({
-            id: 2,
-            slug: "cd34e",
-            name: "Event B",
-            attendee_count: 50,
-            max_attendees: 50,
-          }),
-        ),
-      ];
-      const html = ticketPage({ events, slugs: ["ab12c", "cd34e"] });
-      expect(html).toContain('name="quantity_1" value="1"');
-      expect(html).not.toContain("Select Tickets");
-    });
-  });
-
-  describe("adminSettingsPage", () => {
-    const defaultState: import("#templates/admin/settings.tsx").SettingsPageState =
-      {
-        stripeKeyConfigured: false,
-        stripeKeyMode: null,
-        paymentProvider: "",
-        squareTokenConfigured: false,
-        squareSandbox: false,
-        squareWebhookConfigured: false,
-        webhookUrl: "https://example.com/payment/webhook",
-        bookingFee: "0",
-        embedHosts: "",
-        termsAndConditions: "",
-        businessEmail: "",
-        theme: "light",
-        showPublicSite: false,
-        country: "GB",
-        headerImageUrl: "",
-        storageEnabled: false,
-      };
-
-    test("shows square webhook configured message when key is set", () => {
-      const html = adminSettingsPage(TEST_SESSION, {
-        ...defaultState,
-        paymentProvider: "square",
-        squareTokenConfigured: true,
-        squareWebhookConfigured: true,
-      });
-      expect(html).toContain("A webhook signature key is currently configured");
-      expect(html).toContain("Enter a new key below to replace it");
-    });
-
-    test("shows square webhook not configured message when key is not set", () => {
-      const html = adminSettingsPage(TEST_SESSION, {
-        ...defaultState,
-        paymentProvider: "square",
-        squareTokenConfigured: true,
-      });
-      expect(html).toContain("No webhook signature key is configured");
-      expect(html).toContain("Follow the steps above to set one up");
-    });
-
-    test("shows sandbox checkbox checked when sandbox mode enabled", () => {
-      const html = adminSettingsPage(TEST_SESSION, {
-        ...defaultState,
-        paymentProvider: "square",
-        squareTokenConfigured: true,
-        squareSandbox: true,
-      });
-      expect(html).toContain("Sandbox mode");
-      expect(html).toContain('name="square_sandbox"');
-    });
-
-    test("shows settings sub-navigation", () => {
-      const html = adminSettingsPage(TEST_SESSION, defaultState);
-      expect(html).toContain('href="/admin/settings-advanced"');
-      expect(html).toContain('href="/admin/backup"');
-      expect(html).toContain('href="/admin/debug"');
-    });
-  });
-
-  describe("adminAdvancedSettingsPage", () => {
-    const advancedDefaultState: import("#templates/admin/settings-advanced.tsx").AdvancedSettingsPageState =
-      {
-        showPublicApi: false,
-        emailProvider: "",
-        emailApiKeyConfigured: false,
-        emailFromAddress: "",
-        hostEmailLabel: "",
-        businessEmail: "",
-        confirmationTemplates: { subject: "", html: "", text: "" },
-        adminTemplates: { subject: "", html: "", text: "" },
-        bunnyCdnEnabled: false,
-        bunnyDnsEnabled: false,
-        bunnySubdomain: "",
-        bunnyDnsSubdomainSuffix: "",
-        customDomain: "",
-        customDomainLastValidated: "",
-        cdnHostname: "",
-        appleWalletConfigured: false,
-        appleWalletPassTypeId: "",
-        appleWalletTeamId: "",
-        hostAppleWalletLabel: "",
-        googleWalletConfigured: false,
-        googleWalletIssuerId: "",
-        googleWalletServiceAccountEmail: "",
-        hostGoogleWalletLabel: "",
-        subdomainPreview: "",
-        subdomainPreviewFullDomain: "",
-        theme: "light",
-      };
-
-    test("shows email provider selection when configured", () => {
-      const html = adminAdvancedSettingsPage(TEST_SESSION, {
-        ...advancedDefaultState,
-        emailProvider: "resend",
-        emailFromAddress: "from@test.com",
-      });
-      expect(html).toContain('value="resend"');
-      expect(html).toContain("Send Test Email");
-      expect(html).toContain('value="from@test.com"');
-    });
-
-    test("hides test button when no email provider configured", () => {
-      const html = adminAdvancedSettingsPage(
-        TEST_SESSION,
-        advancedDefaultState,
-      );
-      expect(html).not.toContain("Send Test Email");
-    });
-
-    test("uses business email as from address placeholder", () => {
-      const html = adminAdvancedSettingsPage(TEST_SESSION, {
-        ...advancedDefaultState,
-        businessEmail: "biz@example.com",
-      });
-      expect(html).toContain('placeholder="biz@example.com"');
-    });
-
-    test("uses default placeholder when no business email", () => {
-      const html = adminAdvancedSettingsPage(
-        TEST_SESSION,
-        advancedDefaultState,
-      );
-      expect(html).toContain('placeholder="tickets@yourdomain.com"');
-    });
-
-    test("shows host email label when hostEmailLabel is set", () => {
-      const html = adminAdvancedSettingsPage(TEST_SESSION, {
-        ...advancedDefaultState,
-        hostEmailLabel: "Host Resend (noreply@example.com)",
-      });
-      expect(html).toContain("Host Resend (noreply@example.com)");
-    });
-
-    test("shows None disabled when no hostEmailLabel set", () => {
-      const html = adminAdvancedSettingsPage(
-        TEST_SESSION,
-        advancedDefaultState,
-      );
-      expect(html).toContain("None (disabled)");
-    });
-
-    test("shows warning about careful changes", () => {
-      const html = adminAdvancedSettingsPage(
-        TEST_SESSION,
-        advancedDefaultState,
-      );
-      expect(html).toContain("Be careful changing settings on this page");
-    });
-
-    test("shows breadcrumb back to settings", () => {
-      const html = adminAdvancedSettingsPage(
-        TEST_SESSION,
-        advancedDefaultState,
-      );
-      expect(html).toContain('href="/admin/settings"');
-    });
-
-    test("shows subdomain preview confirmation when subdomainPreview is set", () => {
-      const html = adminAdvancedSettingsPage(TEST_SESSION, {
-        ...advancedDefaultState,
-        bunnyDnsEnabled: true,
-        bunnyDnsSubdomainSuffix: ".tickets.example.com",
-        subdomainPreview: "myevent",
-        subdomainPreviewFullDomain: "myevent.tickets.example.com",
-      });
-      expect(html).toContain("myevent.tickets.example.com");
-      expect(html).toContain("is available");
-      expect(html).toContain('name="save"');
-      expect(html).toContain("Confirm registration");
-      expect(html).toContain('value="myevent"');
-    });
-  });
-
-  describe("adminCalendarPage", () => {
-    const calendarAttendee = (
-      overrides: Partial<CalendarAttendeeRow> = {},
-    ): CalendarAttendeeRow => ({
-      ...testAttendee(),
-      eventName: "Daily Event",
-      eventDate: "",
-      eventLocation: "",
-      eventId: 1,
-      date: "2026-03-15",
-      ...overrides,
-    });
-
-    test("renders Calendar title", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("Calendar");
-      expect(html).toContain("Attendees by Date");
-    });
-
-    test("renders date selector dropdown", () => {
-      const dates = [
-        {
-          value: "2026-03-15",
-          label: "Sunday 15 March 2026",
-          hasBookings: true,
-        },
-        {
-          value: "2026-03-16",
-          label: "Monday 16 March 2026",
-          hasBookings: false,
-        },
-      ];
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        dates,
-        "2026-03-10",
-      );
-      expect(html).toContain("Sunday 15 March 2026");
-      expect(html).toContain("Monday 16 March 2026");
-      expect(html).toContain("Select a date");
-    });
-
-    test("disables options for dates without bookings", () => {
-      const dates = [
-        {
-          value: "2026-03-15",
-          label: "Sunday 15 March 2026",
-          hasBookings: false,
-        },
-      ];
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        dates,
-        "2026-03-10",
-      );
-      expect(html).toContain("<option disabled>");
-    });
-
-    test("enables options for dates with bookings", () => {
-      const dates = [
-        {
-          value: "2026-03-15",
-          label: "Sunday 15 March 2026",
-          hasBookings: true,
-        },
-      ];
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        dates,
-        "2026-03-10",
-      );
-      expect(html).toContain(
-        'value="/admin/calendar?date=2026-03-15#attendees"',
-      );
-    });
-
-    test("shows prompt when no date selected", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("Select a date above to view attendees");
-    });
-
-    test("shows no attendees message when date selected but empty", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("No attendees for this date");
-    });
-
-    test("shows formatted date label when date is selected", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("Sunday 15 March 2026");
-    });
-
-    test("renders attendee rows with event name and link", () => {
-      const attendees = [calendarAttendee()];
-      const html = adminCalendarPage(
-        attendees,
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("Daily Event");
-      expect(html).toContain('href="/admin/event/1"');
-      expect(html).toContain("John Doe");
-    });
-
-    test("renders Event column header", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("<th>Event</th>");
-    });
-
-    test("shows CSV export link when date has attendees", () => {
-      const attendees = [calendarAttendee()];
-      const html = adminCalendarPage(
-        attendees,
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain('href="/admin/calendar/export?date=2026-03-15"');
-      expect(html).toContain("Export CSV");
-    });
-
-    test("does not show CSV export when date has no attendees", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).not.toContain("Export CSV");
-    });
-
-    test("does not show CSV export when no date selected", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        [],
-        "2026-03-10",
-      );
-      expect(html).not.toContain("Export CSV");
-    });
-
-    test("includes Calendar link in admin nav", () => {
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain('href="/admin/calendar"');
-    });
-
-    test("renders empty string for attendee without email", () => {
-      const attendees = [calendarAttendee({ email: "" })];
-      const html = adminCalendarPage(
-        attendees,
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("John Doe");
-    });
-
-    test("escapes attendee data", () => {
-      const attendees = [calendarAttendee({ name: "<script>evil()</script>" })];
-      const html = adminCalendarPage(
-        attendees,
-        "localhost",
-        TEST_SESSION,
-        "2026-03-15",
-        [],
-        "2026-03-10",
-      );
-      expect(html).toContain("&lt;script&gt;");
-    });
-
-    test("places Select a date between past and future dates", () => {
-      const dates = [
-        {
-          value: "2026-03-08",
-          label: "Sunday 8 March 2026",
-          hasBookings: true,
-        },
-        {
-          value: "2026-03-09",
-          label: "Monday 9 March 2026",
-          hasBookings: true,
-        },
-        {
-          value: "2026-03-15",
-          label: "Sunday 15 March 2026",
-          hasBookings: true,
-        },
-        {
-          value: "2026-03-16",
-          label: "Monday 16 March 2026",
-          hasBookings: true,
-        },
-      ];
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        dates,
-        "2026-03-10",
-      );
-      const selectMatch = html.match(/<select[^>]*>([\s\S]*?)<\/select>/)!;
-      const optionTexts = [...selectMatch[1]!.matchAll(/>([^<]+)</g)].map(
-        (m) => m[1],
-      );
-      expect(optionTexts).toEqual([
-        "Sunday 8 March 2026",
-        "Monday 9 March 2026",
-        "Select a date",
-        "Sunday 15 March 2026",
-        "Monday 16 March 2026",
-      ]);
-    });
-
-    test("places Select a date at end when all dates are past", () => {
-      const dates = [
-        {
-          value: "2026-03-08",
-          label: "Sunday 8 March 2026",
-          hasBookings: true,
-        },
-        {
-          value: "2026-03-09",
-          label: "Monday 9 March 2026",
-          hasBookings: true,
-        },
-      ];
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        dates,
-        "2026-03-10",
-      );
-      const selectMatch = html.match(/<select[^>]*>([\s\S]*?)<\/select>/)!;
-      const optionTexts = [...selectMatch[1]!.matchAll(/>([^<]+)</g)].map(
-        (m) => m[1],
-      );
-      expect(optionTexts).toEqual([
-        "Sunday 8 March 2026",
-        "Monday 9 March 2026",
-        "Select a date",
-      ]);
-    });
-
-    test("places Select a date at start when all dates are future", () => {
-      const dates = [
-        {
-          value: "2026-03-15",
-          label: "Sunday 15 March 2026",
-          hasBookings: true,
-        },
-        {
-          value: "2026-03-16",
-          label: "Monday 16 March 2026",
-          hasBookings: true,
-        },
-      ];
-      const html = adminCalendarPage(
-        [],
-        "localhost",
-        TEST_SESSION,
-        null,
-        dates,
-        "2026-03-10",
-      );
-      const selectMatch = html.match(/<select[^>]*>([\s\S]*?)<\/select>/)!;
-      const optionTexts = [...selectMatch[1]!.matchAll(/>([^<]+)</g)].map(
-        (m) => m[1],
-      );
-      expect(optionTexts).toEqual([
-        "Select a date",
-        "Sunday 15 March 2026",
-        "Monday 16 March 2026",
-      ]);
-    });
-  });
-
-  describe("generateCalendarCsv", () => {
-    const calendarAttendee = (
-      overrides: Partial<CalendarAttendeeRow> = {},
-    ): CalendarAttendeeRow => ({
-      ...testAttendee(),
-      eventName: "Daily Event",
-      eventDate: "",
-      eventLocation: "",
-      eventId: 1,
-      date: "2026-03-15",
-      ...overrides,
-    });
-
-    test("generates CSV header for empty attendees (no Event Date/Location columns)", () => {
-      const csv = generateCalendarCsv([]);
-      expect(csv).toBe(
-        "Event,Date,Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
-      );
-    });
-
-    test("omits Event Date and Event Location columns when all empty", () => {
-      const attendees = [calendarAttendee()];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[0]).toContain("Event,Date,Name");
-      expect(lines[1]).toMatch(/^Daily Event,2026-03-15,/);
-    });
-
-    test("includes Event Date column when some attendees have event dates", () => {
-      const attendees = [
-        calendarAttendee({ eventDate: "2026-06-15T14:00:00.000Z" }),
-      ];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[0]).toContain("Event,Event Date,Date,Name");
-      expect(lines[1]).toContain("2026-06-15T14:00:00.000Z");
-    });
-
-    test("includes Event Location column when some attendees have event locations", () => {
-      const attendees = [calendarAttendee({ eventLocation: "Village Hall" })];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[0]).toContain("Event,Event Location,Date,Name");
-      expect(lines[1]).toContain("Village Hall");
-    });
-
-    test("includes Date column", () => {
-      const attendees = [calendarAttendee({ date: "2026-03-20" })];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain("2026-03-20");
-    });
-
-    test("escapes event names with commas", () => {
-      const attendees = [calendarAttendee({ eventName: "Event, Special" })];
-      const csv = generateCalendarCsv(attendees);
-      expect(csv).toContain('"Event, Special"');
-    });
-
-    test("includes standard attendee columns", () => {
-      const attendees = [
-        calendarAttendee({
-          created: "2024-01-15T10:30:00Z",
-          quantity: 2,
-          price_paid: "2000",
-          payment_id: "pi_abc",
-          checked_in: true,
-        }),
-      ];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain("John Doe");
-      expect(lines[1]).toContain("john@example.com");
-      expect(lines[1]).toContain(",2,");
-      expect(lines[1]).toContain("20.00");
-      expect(lines[1]).toContain("pi_abc");
-      expect(lines[1]).toContain(",Yes,");
-    });
-
-    test("generates multiple rows", () => {
-      const attendees = [
-        calendarAttendee(),
-        calendarAttendee({
-          id: 2,
-          name: "Jane Smith",
-          eventName: "Other Event",
-        }),
-      ];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      expect(lines).toHaveLength(3);
-      expect(lines[1]).toContain("Daily Event");
-      expect(lines[2]).toContain("Other Event");
-    });
-
-    test("handles null date in calendar row", () => {
-      const attendees = [calendarAttendee({ date: null })];
-      const csv = generateCalendarCsv(attendees);
-      const lines = csv.split("\n");
-      // The date column should be empty when null
-      expect(lines[1]).toMatch(/^Daily Event,,/);
-    });
-  });
-
-  describe("admin nav Calendar link", () => {
-    test("admin dashboard includes Calendar link in nav", () => {
-      const html = adminDashboardPage([], TEST_SESSION);
-      expect(html).toContain('href="/admin/calendar"');
-      expect(html).toContain("Calendar");
-    });
-  });
-
-  describe("adminEventPage event date and location", () => {
-    test("shows Event Date row when event has a date", () => {
-      const event = testEventWithCount({
-        date: "2026-06-15T14:00:00.000Z",
-        attendee_count: 0,
-      });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Event Date");
-      expect(html).toContain("Monday 15 June 2026 at 15:00 GMT+1");
-    });
-
-    test("does not show Event Date row when date is empty", () => {
-      const event = testEventWithCount({ date: "", attendee_count: 0 });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("Event Date");
-    });
-
-    test("shows Location row when event has a location", () => {
-      const event = testEventWithCount({
-        location: "Village Hall",
-        attendee_count: 0,
-      });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("<th>Location</th>");
-      expect(html).toContain("Village Hall");
-    });
-
-    test("does not show Location row when location is empty", () => {
-      const event = testEventWithCount({ location: "", attendee_count: 0 });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).not.toContain("<th>Location</th>");
-    });
-
-    test("shows both Event Date and Location when both are set", () => {
-      const event = testEventWithCount({
-        date: "2026-06-15T14:00:00.000Z",
-        location: "Town Centre",
-        attendee_count: 0,
-      });
-      const html = adminEventPage({
-        event,
-        attendees: [],
-        allowedDomain: "localhost",
-        session: TEST_SESSION,
-      });
-      expect(html).toContain("Event Date");
-      expect(html).toContain("Town Centre");
-    });
-  });
-
-  describe("adminEventPage edit form pre-fills date and location", () => {
-    test("empty date shows no pre-filled value in edit form", () => {
-      const event = testEventWithCount({ date: "", attendee_count: 0 });
-      const html = adminEventEditPage(event, [], TEST_SESSION, undefined);
-      // The date field should render split date and time inputs
-      expect(html).toContain('name="date_date"');
-      expect(html).toContain('name="date_time"');
-    });
-
-    test("non-empty date shows formatted split values in edit form", () => {
-      const event = testEventWithCount({
-        date: "2026-06-15T14:00:00.000Z",
-        attendee_count: 0,
-      });
-      const html = adminEventEditPage(event, [], TEST_SESSION, undefined);
-      // Should contain split date and time values converted to Europe/London (BST = UTC+1)
-      expect(html).toContain('value="2026-06-15"');
-      expect(html).toContain('value="15:00"');
-    });
-
-    test("pre-fills location in edit form", () => {
-      const event = testEventWithCount({
-        location: "Village Hall",
-        attendee_count: 0,
-      });
-      const html = adminEventEditPage(event, [], TEST_SESSION, undefined);
-      expect(html).toContain('value="Village Hall"');
-    });
-  });
-
-  describe("adminEventEditPage max_price field", () => {
-    test("renders max_price field with value when set", () => {
-      const event = testEventWithCount({
-        can_pay_more: true,
-        max_price: 50000,
-        attendee_count: 0,
-      });
-      const html = adminEventEditPage(event, [], TEST_SESSION, undefined);
-      expect(html).toContain('name="max_price"');
-      expect(html).toContain('value="500.00"');
-    });
-
-    test("renders max_price field with 0.00 when zero", () => {
-      const event = testEventWithCount({
-        can_pay_more: true,
-        max_price: 0,
-        attendee_count: 0,
-      });
-      const html = adminEventEditPage(event, [], TEST_SESSION, undefined);
-      expect(html).toContain('name="max_price"');
-      expect(html).toContain('value="0.00"');
-    });
-  });
-
-  describe("generateAttendeesCsv with eventInfo", () => {
-    test("includes Event Date column when eventInfo has non-empty eventDate", () => {
-      const eventInfo: CsvEventInfo = {
-        eventDate: "2026-06-15T14:00:00.000Z",
-        eventLocation: "",
-      };
-      const csv = generateAttendeesCsv([], false, eventInfo);
-      expect(csv).toContain("Event Date,Name");
-      expect(csv).not.toContain("Event Location");
-    });
-
-    test("includes Event Location column when eventInfo has non-empty eventLocation", () => {
-      const eventInfo: CsvEventInfo = {
-        eventDate: "",
-        eventLocation: "Village Hall",
-      };
-      const csv = generateAttendeesCsv([], false, eventInfo);
-      expect(csv).toContain("Event Location,Name");
-      expect(csv).not.toContain("Event Date");
-    });
-
-    test("includes both Event Date and Event Location columns", () => {
-      const eventInfo: CsvEventInfo = {
-        eventDate: "2026-06-15T14:00:00.000Z",
-        eventLocation: "Village Hall",
-      };
-      const csv = generateAttendeesCsv([], false, eventInfo);
-      expect(csv).toContain("Event Date,Event Location,Name");
-    });
-
-    test("includes event date and location values in rows", () => {
-      const eventInfo: CsvEventInfo = {
-        eventDate: "2026-06-15T14:00:00.000Z",
-        eventLocation: "Village Hall",
-      };
-      const attendees = [testAttendee()];
-      const csv = generateAttendeesCsv(attendees, false, eventInfo);
-      const lines = csv.split("\n");
-      expect(lines[1]).toContain(
-        "2026-06-15T14:00:00.000Z,Village Hall,John Doe",
-      );
-    });
-
-    test("omits Event Date and Event Location when eventInfo is undefined", () => {
-      const csv = generateAttendeesCsv([], false);
-      expect(csv).not.toContain("Event Date");
-      expect(csv).not.toContain("Event Location");
-    });
-
-    test("omits Event Date and Event Location when both are empty", () => {
-      const eventInfo: CsvEventInfo = { eventDate: "", eventLocation: "" };
-      const csv = generateAttendeesCsv([], false, eventInfo);
-      expect(csv).not.toContain("Event Date");
-      expect(csv).not.toContain("Event Location");
-    });
-  });
-
-  describe("datetime validation via eventFields date field", () => {
-    const dateField = eventFields.find((f) => f.name === "date")!;
-
-    test("accepts valid datetime value", () => {
-      const result = dateField.validate?.("2026-06-15T14:00");
-      expect(result).toBeNull();
-    });
-
-    test("rejects invalid datetime value", () => {
-      const result = dateField.validate?.("not-a-date");
-      expect(result).toBe("Please enter a valid date and time");
-    });
-  });
-
-  describe("ticketPage event date and location", () => {
-    const renderTicket = (ev: EventWithCount, opts?: { iframe?: boolean }) => {
-      if (opts?.iframe) detectIframeMode("https://example.com/?iframe=true");
-      else detectIframeMode("https://example.com/");
-      return ticketPage({
-        events: [buildTicketEvent(ev)],
-        slugs: [ev.slug],
-        dates: [],
-      });
-    };
-
-    test("shows date on public ticket page when event has date", () => {
-      const event = testEventWithCount({
-        date: "2026-06-15T14:00:00.000Z",
-        attendee_count: 0,
-      });
-      const html = renderTicket(event);
-      expect(html).toContain("<strong>Date:</strong>");
-      expect(html).toContain("Monday 15 June 2026 at 15:00 GMT+1");
-    });
-
-    test("does not show date on public ticket page when date is empty", () => {
-      const event = testEventWithCount({ date: "", attendee_count: 0 });
-      const html = renderTicket(event);
-      expect(html).not.toContain("<strong>Date:</strong>");
-    });
-
-    test("shows location on public ticket page when event has location", () => {
-      const event = testEventWithCount({
-        location: "Village Hall",
-        attendee_count: 0,
-      });
-      const html = renderTicket(event);
-      expect(html).toContain("<strong>Location:</strong>");
-      expect(html).toContain("Village Hall");
-    });
-
-    test("does not show location on public ticket page when location is empty", () => {
-      const event = testEventWithCount({ location: "", attendee_count: 0 });
-      const html = renderTicket(event);
-      expect(html).not.toContain("<strong>Location:</strong>");
-    });
-
-    test("hides date and location in iframe mode", () => {
-      const event = testEventWithCount({
-        date: "2026-06-15T14:00:00.000Z",
-        location: "Village Hall",
-        attendee_count: 0,
-      });
-      const html = renderTicket(event, { iframe: true });
-      expect(html).not.toContain("<strong>Date:</strong>");
-      expect(html).not.toContain("<strong>Location:</strong>");
-    });
-
-    test("shows past event badge for event with date in the past", () => {
-      const event = testEventWithCount({
-        date: "2020-01-15T14:00:00.000Z",
-        attendee_count: 0,
-      });
-      const html = renderTicket(event);
-      expect(html).toContain("badge-alert");
-      expect(html).toContain("ago");
-    });
-
-    test("does not show past event badge for future event", () => {
-      const event = testEventWithCount({
-        date: "2099-06-15T14:00:00.000Z",
-        attendee_count: 0,
-      });
-      const html = renderTicket(event);
-      expect(html).not.toContain("badge-alert");
-    });
-
-    test("does not show past event badge when date is empty", () => {
-      const event = testEventWithCount({ date: "", attendee_count: 0 });
-      const html = renderTicket(event);
-      expect(html).not.toContain("badge-alert");
-    });
-
-    test("past event badge shows singular day for 1 day ago", () => {
-      const yesterday = addDays(todayInTz(settings.timezone), -1);
-      const event = testEventWithCount({
-        date: `${yesterday}T12:00:00.000Z`,
-        attendee_count: 0,
-      });
-      const html = renderTicket(event);
-      expect(html).toContain("1 day ago");
-      expect(html).not.toContain("(1 day ago)");
-    });
-
-    test("past event badge shows plural days for multiple days ago", () => {
-      const threeDaysAgo = addDays(todayInTz(settings.timezone), -3);
-      const event = testEventWithCount({
-        date: `${threeDaysAgo}T12:00:00.000Z`,
-        attendee_count: 0,
-      });
-      const html = renderTicket(event);
-      expect(html).toContain("3 days ago");
-      expect(html).not.toContain("(3 days ago)");
-    });
-  });
-
-  describe("ticketViewPage event date and location", () => {
-    const token = "AABB0011CCDDEEFF";
-
-    test("shows event date when entry has non-empty event date", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ date: "2026-06-15T14:00:00.000Z" }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).toContain("Monday 15 June 2026 at 15:00 GMT+1");
-    });
-
-    test("does not show event date when event has empty date", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ date: "" }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).not.toContain("ticket-card-date");
-    });
-
-    test("shows location when entry has non-empty location", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ location: "Village Hall" }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).toContain("Village Hall");
-    });
-
-    test("does not show location when event has empty location", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ location: "" }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).not.toContain("ticket-card-location");
-    });
-
-    test("shows both event date and location when both are present", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({
-              date: "2026-06-15T14:00:00.000Z",
-              location: "Town Centre",
-            }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).toContain("Monday 15 June 2026 at 15:00 GMT+1");
-      expect(html).toContain("Town Centre");
-    });
-
-    test("shows each ticket as separate card with SVG endpoint reference", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({
-              id: 1,
-              date: "2026-06-15T14:00:00.000Z",
-            }),
-            attendee: testAttendee({ id: 1 }),
-          },
-          token: "AABB0011CCDDEEF1",
-        },
-        {
-          entry: {
-            event: testEventWithCount({ id: 2, date: "" }),
-            attendee: testAttendee({ id: 2 }),
-          },
-          token: "AABB0011CCDDEEF2",
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).toContain("/t/AABB0011CCDDEEF1/svg");
-      expect(html).toContain("/t/AABB0011CCDDEEF2/svg");
-      expect(html).toContain("2 Tickets");
-    });
-
-    test("hides QR code and token for purchase_only events", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ purchase_only: true }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).not.toContain("ticket-card-qr");
-      expect(html).not.toContain("ticket-card-token");
-      expect(html).not.toContain("/svg");
-    });
-
-    test("hides wallet links for purchase_only events", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ purchase_only: true }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards, true, true);
-      expect(html).not.toContain("wallet-link");
-      expect(html).not.toContain("Apple Wallet");
-      expect(html).not.toContain("Google Wallet");
-    });
-
-    test("hides non-transferable notice for purchase_only events", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({
-              purchase_only: true,
-              non_transferable: true,
-            }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).not.toContain("Non-transferable");
-    });
-
-    test("shows Your Purchase heading for purchase_only events", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ purchase_only: true }),
-            attendee: testAttendee(),
-          },
-          token,
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).toContain("Your Purchase");
-      expect(html).not.toContain("Ticket");
-    });
-
-    test("shows ticket count heading for mixed events", () => {
-      const cards = [
-        {
-          entry: {
-            event: testEventWithCount({ id: 1, purchase_only: true }),
-            attendee: testAttendee({ id: 1 }),
-          },
-          token: "AABB0011CCDDEEF1",
-        },
-        {
-          entry: {
-            event: testEventWithCount({ id: 2, purchase_only: false }),
-            attendee: testAttendee({ id: 2 }),
-          },
-          token: "AABB0011CCDDEEF2",
-        },
-      ];
-      const html = ticketViewPage(cards);
-      expect(html).toContain("2 Tickets");
-      expect(html).not.toContain("Your Purchase");
-    });
-  });
-
-  describeWithEnv(
-    "event images",
-    { env: { STORAGE_ZONE_NAME: "testzone", STORAGE_ZONE_KEY: "testkey" } },
-    () => {
-      describe("renderEventImage", () => {
-        test("returns empty string when image_url is null", () => {
-          const html = renderEventImage({ image_url: "" });
-          expect(html).toBe("");
-        });
-
-        test("renders img tag with proxy URL when image_url is set", () => {
-          const html = renderEventImage({
-            image_url: "abc123.jpg",
-          });
-          expect(html).toContain("/image/abc123.jpg");
-          expect(html).toContain('alt=""');
-          expect(html).toContain('class="event-image"');
-        });
-
-        test("uses empty alt text for decorative image", () => {
-          const html = renderEventImage({
-            image_url: "img.jpg",
-          });
-          expect(html).toContain('alt=""');
-        });
-      });
-
-      describe("ticketPage with image", () => {
-        const renderSingleEvent = (ev: EventWithCount) =>
-          ticketPage({
-            events: [buildTicketEvent(ev)],
-            slugs: [ev.slug],
-            dates: [],
-            terms: null,
-          });
-
-        test("shows event image when image_url is set", () => {
-          const event = testEventWithCount({ image_url: "event-img.jpg" });
-          const html = renderSingleEvent(event);
-          expect(html).toContain("/image/event-img.jpg");
-          expect(html).toContain('class="event-image"');
-        });
-
-        test("does not show image when image_url is null", () => {
-          const event = testEventWithCount({ image_url: "" });
-          const html = renderSingleEvent(event);
-          expect(html).not.toContain("/image/");
-        });
-
-        test("does not show image in iframe mode", () => {
-          detectIframeMode("https://example.com/?iframe=true");
-          const event = testEventWithCount({ image_url: "event-img.jpg" });
-          const html = renderSingleEvent(event);
-          expect(html).not.toContain("event-img.jpg");
-          detectIframeMode("https://example.com/");
-        });
-      });
-
-      describe("ticketPage with images", () => {
-        test("shows image before each event with image_url", () => {
-          const events = [
-            buildTicketEvent(
-              testEventWithCount({
-                id: 1,
-                name: "Event A",
-                image_url: "img-a.jpg",
-              }),
-            ),
-            buildTicketEvent(
-              testEventWithCount({
-                id: 2,
-                name: "Event B",
-                image_url: "img-b.jpg",
-              }),
-            ),
-          ];
-          const html = ticketPage({ events, slugs: ["slug-a", "slug-b"] });
-          expect(html).toContain("/image/img-a.jpg");
-          expect(html).toContain("/image/img-b.jpg");
-        });
-
-        test("does not show images when image_url is null", () => {
-          const events = [
-            buildTicketEvent(
-              testEventWithCount({ id: 1, name: "Event A", image_url: "" }),
-            ),
-          ];
-          const html = ticketPage({ events, slugs: ["slug-a"] });
-          expect(html).not.toContain("/image/");
-        });
-      });
-
-      describe("ticketViewPage ticket count", () => {
-        const token = "AABB0011CCDDEEFF";
-
-        test("shows '1 Ticket' for single ticket", () => {
-          const cards = [
-            {
-              entry: {
-                event: testEventWithCount({ id: 1 }),
-                attendee: testAttendee({ id: 1 }),
-              },
-              token,
-            },
-          ];
-          const html = ticketViewPage(cards);
-          expect(html).toContain("1 Ticket");
-        });
-
-        test("shows '2 Tickets' for multiple tickets", () => {
-          const cards = [
-            {
-              entry: {
-                event: testEventWithCount({ id: 1 }),
-                attendee: testAttendee({ id: 1 }),
-              },
-              token: "AABB0011CCDDEEF1",
-            },
-            {
-              entry: {
-                event: testEventWithCount({ id: 2 }),
-                attendee: testAttendee({ id: 2 }),
-              },
-              token: "AABB0011CCDDEEF2",
-            },
-          ];
-          const html = ticketViewPage(cards);
-          expect(html).toContain("2 Tickets");
-        });
-      });
-
-      describe("ticketViewPage with image", () => {
-        const token = "AABB0011CCDDEEFF";
-
-        test("shows image when event has image_url", () => {
-          const cards = [
-            {
-              entry: {
-                event: testEventWithCount({ image_url: "ticket-img.jpg" }),
-                attendee: testAttendee(),
-              },
-              token,
-            },
-          ];
-          const html = ticketViewPage(cards);
-          expect(html).toContain("/image/ticket-img.jpg");
-          expect(html).toContain('class="ticket-card-image"');
-        });
-
-        test("does not show image when image_url is empty", () => {
-          const cards = [
-            {
-              entry: {
-                event: testEventWithCount({ image_url: "" }),
-                attendee: testAttendee(),
-              },
-              token,
-            },
-          ];
-          const html = ticketViewPage(cards);
-          expect(html).not.toContain("ticket-card-image");
-        });
-      });
-
-      describe("adminDashboardPage with images", () => {
-        test("shows thumbnail when event has image_url", () => {
-          const events = [testEventWithCount({ image_url: "thumb.jpg" })];
-          const html = adminDashboardPage(events, TEST_SESSION);
-          expect(html).toContain("/image/thumb.jpg");
-          expect(html).toContain('class="event-thumbnail"');
-        });
-
-        test("does not show thumbnail when event has no image_url", () => {
-          const events = [testEventWithCount({ image_url: "" })];
-          const html = adminDashboardPage(events, TEST_SESSION);
-          expect(html).not.toContain('src="/image/');
-        });
-      });
-
-      describe("adminEventPage image section", () => {
-        test("does not show image upload on detail page", () => {
-          const event = testEventWithCount({ image_url: "" });
-          const html = adminEventPage({
-            event,
-            attendees: [],
-            allowedDomain: "localhost",
-            session: TEST_SESSION,
-          });
-          expect(html).not.toContain('type="file"');
-          expect(html).not.toContain('name="image"');
-        });
-      });
-
-      describe("adminEventEditPage image section", () => {
-        test("shows image upload field when storage enabled", () => {
-          const event = testEventWithCount({ image_url: "" });
-          const html = adminEventEditPage(event, [], TEST_SESSION);
-          expect(html).toContain('type="file"');
-          expect(html).toContain('name="image"');
-          expect(html).toContain("multipart/form-data");
-        });
-
-        test("shows current image and remove button when image is set", () => {
-          runWithStorageConfig(
-            { zoneName: "testzone", zoneKey: "testkey" },
-            () => {
-              const event = testEventWithCount({ image_url: "current.jpg" });
-              const html = adminEventEditPage(event, [], TEST_SESSION);
-              expect(html).toContain("/image/current.jpg");
-              expect(html).toContain("Remove Image");
-              expect(html).toContain("/image/delete");
-            },
-          );
-        });
-
-        test("does not show image field when storage is not enabled", () => {
-          withStorageDisabled(() => {
-            const event = testEventWithCount({ image_url: "" });
-            const html = adminEventEditPage(event, [], TEST_SESSION);
-            expect(html).not.toContain('type="file"');
-            expect(html).not.toContain('name="image"');
-          });
-        });
-
-        test("shows full-width image preview when event has image", () => {
-          runWithStorageConfig(
-            { zoneName: "testzone", zoneKey: "testkey" },
-            () => {
-              const event = testEventWithCount({ image_url: "preview.jpg" });
-              const html = adminEventEditPage(event, [], TEST_SESSION);
-              expect(html).toContain("event-image-full");
-              expect(html).toContain("/image/preview.jpg");
-            },
-          );
-        });
-      });
-
-      describe("adminDuplicateEventPage image section", () => {
-        test("does not show image upload field even when storage enabled", () => {
-          const event = testEventWithCount({ image_url: "" });
-          const html = adminDuplicateEventPage(event, [], TEST_SESSION);
-          expect(html).not.toContain('type="file"');
-          expect(html).not.toContain('name="image"');
-          expect(html).toContain("multipart/form-data");
-        });
-
-        test("does not show image field when storage is not enabled", () => {
-          withStorageDisabled(() => {
-            const event = testEventWithCount({ image_url: "" });
-            const html = adminDuplicateEventPage(event, [], TEST_SESSION);
-            expect(html).not.toContain('type="file"');
-            expect(html).not.toContain('name="image"');
-          });
-        });
-      });
-
-      describe("adminEventNewPage image section", () => {
-        test("shows image upload field on create form when storage enabled", () => {
-          runWithStorageConfig(
-            { zoneName: "testzone", zoneKey: "testkey" },
-            () => {
-              const html = adminEventNewPage([], TEST_SESSION);
-              expect(html).toContain('type="file"');
-              expect(html).toContain('name="image"');
-              expect(html).toContain("multipart/form-data");
-            },
-          );
-        });
-
-        test("does not show image field on create form when storage is not enabled", () => {
-          withStorageDisabled(() => {
-            const html = adminEventNewPage([], TEST_SESSION);
-            expect(html).not.toContain('type="file"');
-          });
-        });
-      });
-    },
-  );
-
-  describe("adminQuestionsPage", () => {
-    test("renders empty state when no questions", () => {
-      const html = adminQuestionsPage([], TEST_SESSION);
-      expect(html).toContain("No custom questions yet");
-    });
-
-    test("renders question list with answer counts", () => {
-      const html = adminQuestionsPage(
-        [
-          {
-            id: 1,
-            text: "Favourite colour?",
-            answers: [
-              { id: 10, question_id: 1, text: "Red", sort_order: 0 },
-              { id: 11, question_id: 1, text: "Blue", sort_order: 1 },
-            ],
-          },
-        ],
-        TEST_SESSION,
-      );
-      expect(html).toContain("Favourite colour?");
-      expect(html).toContain("2 answers");
-    });
-
-    test("renders singular answer count for one answer", () => {
-      const html = adminQuestionsPage(
-        [
-          {
-            id: 1,
-            text: "Yes or no?",
-            answers: [{ id: 10, question_id: 1, text: "Yes", sort_order: 0 }],
-          },
-        ],
-        TEST_SESSION,
-      );
-      expect(html).toContain("1 answer)");
-      expect(html).not.toContain("1 answers");
-    });
-
-    test("renders error message when provided", () => {
-      const html = adminQuestionsPage([], TEST_SESSION, "Something went wrong");
-      expect(html).toContain("Something went wrong");
-    });
-  });
-
-  describe("adminQuestionPage", () => {
-    const question = {
-      id: 1,
-      text: "T-shirt size?",
-      answers: [
-        { id: 10, question_id: 1, text: "Small", sort_order: 0 },
-        { id: 11, question_id: 1, text: "Large", sort_order: 1 },
-      ],
-    };
-
-    test("renders question text and edit form", () => {
-      const html = adminQuestionPage(question, TEST_SESSION);
-      expect(html).toContain("T-shirt size?");
-      expect(html).toContain('action="/admin/questions/1/edit"');
-    });
-
-    test("renders answer list with delete links", () => {
-      const html = adminQuestionPage(question, TEST_SESSION);
-      expect(html).toContain("Small");
-      expect(html).toContain("Large");
-      expect(html).toContain("/admin/questions/1/answers/10/delete");
-      expect(html).toContain("/admin/questions/1/answers/11/delete");
-    });
-
-    test("renders delete question link", () => {
-      const html = adminQuestionPage(question, TEST_SESSION);
-      expect(html).toContain('href="/admin/questions/1/delete"');
-    });
-
-    test("renders error message when provided", () => {
-      const html = adminQuestionPage(question, TEST_SESSION, "Error!");
-      expect(html).toContain("Error!");
-    });
-
-    test("renders empty answers state", () => {
-      const html = adminQuestionPage(
-        { id: 1, text: "Q?", answers: [] },
-        TEST_SESSION,
-      );
-      expect(html).toContain("No answers yet");
-    });
-
-    test("renders answer counts when provided", () => {
-      const counts = new Map([
-        [10, 5],
-        [11, 3],
-      ]);
-      const html = adminQuestionPage(question, TEST_SESSION, undefined, counts);
-      expect(html).toContain("(5)");
-      expect(html).toContain("(3)");
-    });
-
-    test("renders move-up and move-down buttons", () => {
-      const html = adminQuestionPage(question, TEST_SESSION);
-      expect(html).toContain("/answers/10/move-down");
-      expect(html).not.toContain("/answers/10/move-up");
-      expect(html).toContain("/answers/11/move-up");
-      expect(html).not.toContain("/answers/11/move-down");
-    });
-
-    test("renders both move buttons for middle answer", () => {
-      const q = {
-        id: 1,
-        text: "Q?",
-        answers: [
-          { id: 10, question_id: 1, text: "A", sort_order: 0 },
-          { id: 11, question_id: 1, text: "B", sort_order: 1 },
-          { id: 12, question_id: 1, text: "C", sort_order: 2 },
-        ],
-      };
-      const html = adminQuestionPage(q, TEST_SESSION);
-      expect(html).toContain("/answers/11/move-up");
-      expect(html).toContain("/answers/11/move-down");
-    });
-  });
-
-  describe("adminQuestionDeletePage", () => {
-    const question = {
-      id: 1,
-      text: "T-shirt size?",
-      answers: [{ id: 10, question_id: 1, text: "Small", sort_order: 0 }],
-    };
-
-    test("renders confirmation form with question text", () => {
-      const html = adminQuestionDeletePage(question, TEST_SESSION);
-      expect(html).toContain("Delete Question");
-      expect(html).toContain("T-shirt size?");
-      expect(html).toContain('name="confirm_identifier"');
-      expect(html).toContain('action="/admin/questions/1/delete"');
-    });
-
-    test("warns about cascading deletes", () => {
-      const html = adminQuestionDeletePage(question, TEST_SESSION);
-      expect(html).toContain("all its answers");
-      expect(html).toContain("attendee responses");
-    });
-
-    test("renders error message when provided", () => {
-      const html = adminQuestionDeletePage(
-        question,
-        TEST_SESSION,
-        "Text does not match",
-      );
-      expect(html).toContain("Text does not match");
-    });
-  });
-
-  describe("adminAnswerDeletePage", () => {
-    const question = {
-      id: 1,
-      text: "T-shirt size?",
-      answers: [
-        { id: 10, question_id: 1, text: "Small", sort_order: 0 },
-        { id: 11, question_id: 1, text: "Large", sort_order: 1 },
-      ],
-    };
-    const answer = question.answers[0]!;
-
-    test("renders confirmation form with answer text", () => {
-      const html = adminAnswerDeletePage(question, answer, TEST_SESSION);
-      expect(html).toContain("Delete Answer");
-      expect(html).toContain("Small");
-      expect(html).toContain('name="confirm_identifier"');
-      expect(html).toContain('action="/admin/questions/1/answers/10/delete"');
-    });
-
-    test("shows question context", () => {
-      const html = adminAnswerDeletePage(question, answer, TEST_SESSION);
-      expect(html).toContain("T-shirt size?");
-    });
-
-    test("renders error message when provided", () => {
-      const html = adminAnswerDeletePage(
-        question,
-        answer,
-        TEST_SESSION,
-        "Text does not match",
-      );
-      expect(html).toContain("Text does not match");
-    });
-  });
-
-  describe("adminEventQuestionsPage", () => {
-    test("shows empty state when no questions exist", () => {
-      const event = testEventWithCount({ id: 1, name: "My Event" });
-      const html = adminEventQuestionsPage(event, [], new Set(), TEST_SESSION);
-      expect(html).toContain("No questions created yet");
-      expect(html).toContain('href="/admin/questions"');
-      expect(html).toContain("Create questions");
-    });
-
-    test("shows singular option count for question with one answer", () => {
-      const event = testEventWithCount({ id: 1, name: "My Event" });
-      const questions = [
-        {
-          id: 1,
-          text: "Yes or no?",
-          answers: [{ id: 10, question_id: 1, text: "Yes", sort_order: 0 }],
-        },
-      ];
-      const html = adminEventQuestionsPage(
-        event,
-        questions,
-        new Set(),
-        TEST_SESSION,
-      );
-      expect(html).toContain("1 option: Yes)");
-      expect(html).not.toContain("1 options");
-    });
-
-    test("shows Manage Questions link below form", () => {
-      const event = testEventWithCount({ id: 1, name: "My Event" });
-      const questions = [
-        {
-          id: 1,
-          text: "Q?",
-          answers: [
-            { id: 10, question_id: 1, text: "A", sort_order: 0 },
-            { id: 11, question_id: 1, text: "B", sort_order: 1 },
-          ],
-        },
-      ];
-      const html = adminEventQuestionsPage(
-        event,
-        questions,
-        new Set(),
-        TEST_SESSION,
-      );
-      expect(html).toContain('href="/admin/questions"');
-      expect(html).toContain("Manage Questions");
-    });
-
-    test("lists option names in parentheses", () => {
-      const event = testEventWithCount({ id: 1, name: "My Event" });
-      const questions = [
-        {
-          id: 1,
-          text: "Size?",
-          answers: [
-            { id: 10, question_id: 1, text: "S", sort_order: 0 },
-            { id: 11, question_id: 1, text: "M", sort_order: 1 },
-            { id: 12, question_id: 1, text: "L", sort_order: 2 },
-          ],
-        },
-      ];
-      const html = adminEventQuestionsPage(
-        event,
-        questions,
-        new Set(),
-        TEST_SESSION,
-      );
-      expect(html).toContain("3 options: S, M, L)");
-    });
-  });
-
-  describe("buildAnswerSummaryRows", () => {
-    test("returns empty string when questionData is undefined", () => {
-      expect(buildAnswerSummaryRows(undefined)).toBe("");
-    });
-
-    test("returns empty string when no questions", () => {
-      expect(
-        buildAnswerSummaryRows({ questions: [], attendeeAnswerMap: new Map() }),
-      ).toBe("");
-    });
-
-    test("renders question with answer counts", () => {
-      const html = buildAnswerSummaryRows({
-        questions: [
-          {
-            id: 1,
-            text: "Size?",
-            answers: [
-              { id: 10, question_id: 1, text: "Small", sort_order: 0 },
-              { id: 11, question_id: 1, text: "Large", sort_order: 1 },
-            ],
-          },
-        ],
-        attendeeAnswerMap: new Map([
-          [1, [10]],
-          [2, [10]],
-          [3, [11]],
-        ]),
-      });
-      expect(html).toContain("<th>Size?</th>");
-      expect(html).toContain("Small (2)");
-      expect(html).toContain("Large (1)");
-    });
-
-    test("shows zero for answers with no selections", () => {
-      const html = buildAnswerSummaryRows({
-        questions: [
-          {
-            id: 1,
-            text: "Q?",
-            answers: [{ id: 10, question_id: 1, text: "A", sort_order: 0 }],
-          },
-        ],
-        attendeeAnswerMap: new Map(),
-      });
-      expect(html).toContain("A (0)");
-    });
-  });
-
-  describe("adminEventPage with questionData", () => {
-    test("renders answer summary rows in details table", () => {
-      const html = adminEventPage({
-        event: testEventWithCount({ id: 1, name: "E" }),
-        attendees: [],
-        allowedDomain: "example.com",
-        session: TEST_SESSION,
-        questionData: {
-          questions: [
-            {
-              id: 1,
-              text: "Size?",
-              answers: [{ id: 10, question_id: 1, text: "S", sort_order: 0 }],
-            },
-          ],
-          attendeeAnswerMap: new Map(),
-        },
-      });
-      expect(html).toContain("<th>Size?</th>");
-      expect(html).toContain("S (0)");
-    });
-  });
-});
-
-describeWithEnv(
-  "read-only mode templates",
-  { env: { READ_ONLY: "true" } },
-  () => {
-    test("AdminNav shows read-only banner", () => {
-      const html = String(
-        AdminNav({ session: TEST_SESSION, active: "/admin/" }),
-      );
-      expect(html).toContain("read-only-banner");
-      expect(html).toContain("This site is in read-only mode");
-    });
-
-    test("ticketPage hides booking form in read-only mode", () => {
-      const event = testEventWithCount({ attendee_count: 0 });
-      const html = ticketPage({
-        events: [buildTicketEvent(event)],
-        slugs: [event.slug],
-        dates: [],
-      });
-      expect(html).toContain("Registration closed.");
-      expect(html).not.toContain("Continue");
-    });
-  },
-);
