+++ test/templates/csv.test.ts
@@ -1,0 +1,239 @@
+import { expect } from "@std/expect";
+import { beforeAll, describe, it as test } from "@std/testing/bdd";
+import { signCsrfToken } from "#lib/csrf.ts";
+import { type CsvEventInfo, generateAttendeesCsv } from "#templates/csv.ts";
+import { setupTestEncryptionKey, testAttendee } from "#test-utils";
+
+beforeAll(async () => {
+  setupTestEncryptionKey();
+  await signCsrfToken();
+});
+
+describe("generateAttendeesCsv", () => {
+  test("generates CSV header for empty attendees", () => {
+    const csv = generateAttendeesCsv([]);
+    expect(csv).toBe(
+      "Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
+    );
+  });
+
+  test("generates CSV with attendee data", () => {
+    const attendees = [
+      testAttendee({ created: "2024-01-15T10:30:00Z", quantity: 2 }),
+    ];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[0]).toBe(
+      "Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
+    );
+    expect(lines[1]).toContain("John Doe");
+    expect(lines[1]).toContain("john@example.com");
+    expect(lines[1]).toContain(",2,");
+    expect(lines[1]).toContain("2024-01-15T10:30:00.000Z");
+  });
+
+  test("escapes values with commas", () => {
+    const attendees = [testAttendee({ name: "Doe, John" })];
+    const csv = generateAttendeesCsv(attendees);
+    expect(csv).toContain('"Doe, John"');
+  });
+
+  test("escapes values with quotes", () => {
+    const attendees = [testAttendee({ name: 'John "JD" Doe' })];
+    const csv = generateAttendeesCsv(attendees);
+    expect(csv).toContain('"John ""JD"" Doe"');
+  });
+
+  test("escapes values with newlines", () => {
+    const attendees = [testAttendee({ name: "John\nDoe" })];
+    const csv = generateAttendeesCsv(attendees);
+    expect(csv).toContain('"John\nDoe"');
+  });
+
+  test("generates multiple rows", () => {
+    const attendees = [
+      testAttendee(),
+      testAttendee({
+        id: 2,
+        name: "Jane Smith",
+        email: "jane@example.com",
+        created: "2024-01-16T11:00:00Z",
+        quantity: 3,
+      }),
+    ];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines).toHaveLength(3);
+    expect(lines[1]).toContain("John Doe");
+    expect(lines[2]).toContain("Jane Smith");
+  });
+
+  test("includes phone number in CSV output", () => {
+    const attendees = [testAttendee({ phone: "+1 555 123 4567" })];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[0]).toBe(
+      "Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
+    );
+    expect(lines[1]).toContain("+1 555 123 4567");
+  });
+
+  test("includes empty phone column when phone not collected", () => {
+    const attendees = [testAttendee()];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain("john@example.com,,,,1,");
+  });
+
+  test("generates CSV with price and transaction ID", () => {
+    const attendees = [
+      testAttendee({
+        payment_id: "pi_abc123",
+        quantity: 2,
+        price_paid: "2000",
+      }),
+    ];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain("20.00");
+    expect(lines[1]).toContain("pi_abc123");
+  });
+
+  test("formats zero price and empty payment_id for free attendees", () => {
+    const attendees = [testAttendee()];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    // Price should be 0.00, transaction ID empty, followed by Checked In, Token, URL
+    expect(lines[1]).toContain("0.00,,No,");
+  });
+
+  test("shared transaction ID across multiple attendees", () => {
+    const attendees = [
+      testAttendee({
+        payment_id: "pi_shared_123",
+        price_paid: "1000",
+      }),
+      testAttendee({
+        id: 2,
+        event_id: 2,
+        payment_id: "pi_shared_123",
+        quantity: 2,
+        price_paid: "3000",
+      }),
+    ];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain("10.00");
+    expect(lines[1]).toContain("pi_shared_123");
+    expect(lines[2]).toContain("30.00");
+    expect(lines[2]).toContain("pi_shared_123");
+  });
+
+  test("includes Checked In as Yes for checked-in attendee", () => {
+    const attendees = [testAttendee({ checked_in: true })];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain(",Yes,");
+  });
+
+  test("includes Checked In as No for not checked-in attendee", () => {
+    const attendees = [testAttendee({ checked_in: false })];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain(",No,");
+  });
+
+  test("includes ticket token and URL in CSV output", () => {
+    const attendees = [testAttendee({ ticket_token: "abc123" })];
+    const csv = generateAttendeesCsv(attendees);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain("abc123");
+    expect(lines[1]).toContain("https://localhost/t/abc123");
+  });
+
+  test("includes Date column when includeDate is true", () => {
+    const csv = generateAttendeesCsv([], true);
+    expect(csv).toBe(
+      "Date,Name,Email,Phone,Address,Special Instructions,Quantity,Registered,Price Paid,Transaction ID,Checked In,Ticket Token,Ticket URL",
+    );
+  });
+
+  test("includes date value in row when includeDate is true", () => {
+    const attendees = [testAttendee({ date: "2026-03-15" })];
+    const csv = generateAttendeesCsv(attendees, true);
+    const lines = csv.split("\n");
+    expect(lines[0]).toContain("Date,Name");
+    expect(lines[1]).toMatch(/^2026-03-15,/);
+  });
+
+  test("includes empty date in row when date is null", () => {
+    const attendees = [testAttendee({ date: null })];
+    const csv = generateAttendeesCsv(attendees, true);
+    const lines = csv.split("\n");
+    expect(lines[1]).toMatch(/^,John Doe/);
+  });
+
+  test("omits Date column when includeDate is false", () => {
+    const attendees = [testAttendee({ date: "2026-03-15" })];
+    const csv = generateAttendeesCsv(attendees, false);
+    expect(csv.startsWith("Name,")).toBe(true);
+    expect(csv).not.toContain("2026-03-15");
+  });
+});
+
+describe("generateAttendeesCsv with eventInfo", () => {
+  test("includes Event Date column when eventInfo has non-empty eventDate", () => {
+    const eventInfo: CsvEventInfo = {
+      eventDate: "2026-06-15T14:00:00.000Z",
+      eventLocation: "",
+    };
+    const csv = generateAttendeesCsv([], false, eventInfo);
+    expect(csv).toContain("Event Date,Name");
+    expect(csv).not.toContain("Event Location");
+  });
+
+  test("includes Event Location column when eventInfo has non-empty eventLocation", () => {
+    const eventInfo: CsvEventInfo = {
+      eventDate: "",
+      eventLocation: "Village Hall",
+    };
+    const csv = generateAttendeesCsv([], false, eventInfo);
+    expect(csv).toContain("Event Location,Name");
+    expect(csv).not.toContain("Event Date");
+  });
+
+  test("includes both Event Date and Event Location columns", () => {
+    const eventInfo: CsvEventInfo = {
+      eventDate: "2026-06-15T14:00:00.000Z",
+      eventLocation: "Village Hall",
+    };
+    const csv = generateAttendeesCsv([], false, eventInfo);
+    expect(csv).toContain("Event Date,Event Location,Name");
+  });
+
+  test("includes event date and location values in rows", () => {
+    const eventInfo: CsvEventInfo = {
+      eventDate: "2026-06-15T14:00:00.000Z",
+      eventLocation: "Village Hall",
+    };
+    const attendees = [testAttendee()];
+    const csv = generateAttendeesCsv(attendees, false, eventInfo);
+    const lines = csv.split("\n");
+    expect(lines[1]).toContain(
+      "2026-06-15T14:00:00.000Z,Village Hall,John Doe",
+    );
+  });
+
+  test("omits Event Date and Event Location when eventInfo is undefined", () => {
+    const csv = generateAttendeesCsv([], false);
+    expect(csv).not.toContain("Event Date");
+    expect(csv).not.toContain("Event Location");
+  });
+
+  test("omits Event Date and Event Location when both are empty", () => {
+    const eventInfo: CsvEventInfo = { eventDate: "", eventLocation: "" };
+    const csv = generateAttendeesCsv([], false, eventInfo);
+    expect(csv).not.toContain("Event Date");
+    expect(csv).not.toContain("Event Location");
+  });
+});
